% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{active_samples}
\alias{active_samples}
\alias{active_samples.default}
\alias{active_samples.FacileDataSet}
\alias{active_samples<-}
\alias{active_samples<-.FacileDataSet}
\title{Manipulate / query the "active samples" from a FacileDataStore}
\usage{
active_samples(x, ...)

\method{active_samples}{default}(x, ....)

\method{active_samples}{FacileDataSet}(x, ...)

active_samples(x) <- value

\method{active_samples}{FacileDataSet}(x) <- value
}
\arguments{
\item{x}{A `FacileDataStore`}

\item{value}{a sample descriptor}
}
\description{
A FacileDataStore can house an enormous set of samples, however we can
configure it to work only on a subset of them at a time. The subset of
samples currently under scrutiny are the "active samples"
}
