% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.FacileDataSet.R
\name{adata}
\alias{adata}
\title{Bioc-container specific assay data extraction functions}
\usage{
adata(x, assay = NULL, ...)
}
\arguments{
\item{x}{SummarizedExperiment, ExpressionSet or DGEList}

\item{...}{additional args, ignored for now}
}
\description{
Get assay matrix
}
