% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construction.R
\name{addFacileSampleCovariates}
\alias{addFacileSampleCovariates}
\title{A data.frame of dataset,sample_id, ... to map covariates to samples}
\usage{
addFacileSampleCovariates(x)
}
\description{
A data.frame of dataset,sample_id, ... to map covariates to samples
}
