% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{an_fds}
\alias{an_fds}
\title{Retrieves a multi-modal FacileDataSet based on KPMP data.}
\usage{
an_fds()
}
\value{
a FacileDataSet
}
\description{
We use this dataset for examples and unit testing purposes.
}
