% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{an_se_list}
\alias{an_se_list}
\title{Creates a list of SummarizedExperiment to turn into an_fds FDS}
\usage{
an_se_list()
}
\value{
list of names SummarizedExperiemnts
}
\description{
Creates a list of SummarizedExperiment to turn into an_fds FDS
}
