% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.BiocAssayContainers.R
\name{as.DGEList}
\alias{as.DGEList}
\alias{as.DGEList.data.frame}
\alias{as.DGEList.tbl}
\alias{as.DGEList.facile_frame}
\alias{as.DGEList.FacileDataSet}
\alias{as.ExpressionSet}
\alias{as.ExpressionSet.data.frame}
\alias{as.ExpressionSet.FacileDataSet}
\alias{as.SummarizedExperiment}
\alias{as.SummarizedExperiment.data.frame}
\alias{as.SummarizedExperiment.FacileDataSet}
\title{Converts a "facile result" to a traditional Bioconductor assay container.}
\usage{
as.DGEList(x, ...)

\method{as.DGEList}{data.frame}(
  x,
  covariates = TRUE,
  feature_ids = NULL,
  assay_name = NULL,
  .fds = NULL,
  custom_key = Sys.getenv("USER"),
  ...
)

\method{as.DGEList}{tbl}(
  x,
  covariates = TRUE,
  feature_ids = NULL,
  assay_name = NULL,
  .fds = NULL,
  custom_key = Sys.getenv("USER"),
  ...
)

\method{as.DGEList}{facile_frame}(
  x,
  covariates = TRUE,
  feature_ids = NULL,
  assay_name = NULL,
  custom_key = Sys.getenv("USER"),
  ...
)

\method{as.DGEList}{FacileDataSet}(
  x,
  covariates = TRUE,
  feature_ids = NULL,
  assay_name = NULL,
  custom_key = Sys.getenv("USER"),
  ...
)

as.ExpressionSet(x, ...)

\method{as.ExpressionSet}{data.frame}(
  x,
  covariates = TRUE,
  feature_ids = NULL,
  assay_name = default_assay(.fds),
  .fds = fds(x),
  custom_key = Sys.getenv("USER"),
  ...
)

\method{as.ExpressionSet}{FacileDataSet}(
  x,
  covariates = TRUE,
  feature_ids = NULL,
  assay_name = default_assay(.fds),
  .fds = fds(x),
  custom_key = Sys.getenv("USER"),
  ...
)

as.SummarizedExperiment(x, ...)

\method{as.SummarizedExperiment}{data.frame}(
  x,
  covariates = TRUE,
  feature_ids = NULL,
  assay_name = default_assay(.fds),
  .fds = fds(x),
  custom_key = Sys.getenv("USER"),
  ...
)

\method{as.SummarizedExperiment}{FacileDataSet}(
  x,
  covariates = TRUE,
  feature_ids = NULL,
  assay_name = default_assay(.fds),
  .fds = fds(x),
  custom_key = Sys.getenv("USER"),
  ...
)
}
\arguments{
\item{x}{a facile expression-like result}

\item{covariates}{The covariates the user wants to add to the $samples of
the DGEList. This can take the following forms:
\itemize{
\item \code{TRUE}: All covariates are retrieved from the \code{FacileDataSet}
\item \code{FALSE}: TODO: Better handle FALSE
\item \code{character}: A vector of covariate names to fetch from the
\code{FacileDataSet}. Must be elements of \code{names(sample_definitions(x))}
\item \code{data.frame}: A wide covariate table (dataset, sample_id, covariates ...)
This may be external covariates for samples not available within
\code{x} (yet), ie. a table of covariates provided by a third party.
\item \code{NULL}: do not decorate with \emph{any} covariates.
}}

\item{feature_ids}{the features to get expression for (if not specified
in \code{x} descriptor). These correspond to the elements found in the
\code{feature_info_tbl(x)$feature_id} column.}

\item{assay_name}{the name of the assay matrix to use when populating the
default assay matrix of the bioconductor container (the \verb{$counts}
matrix of a \code{DGEList}, the \code{exprs()} of an \code{ExpressionSet}, etc.).
The default value is the entry provided by \code{\link[=default_assay]{default_assay()}}}

\item{.fds}{The \code{FacileDataSet} that \code{x} was retrieved from}

\item{custom_key}{the custom key to use to fetch custom annotations from
\code{.fds}}
}
\value{
the appropriate bioconductor assay container, ie. an \code{edgeR::DGEList}
for \code{as.DGEList}, a \code{Biobase::ExpressionSet} for \code{as.ExpressionSet}, or
a \code{SummarizedExperiment::SummarizedExperiment} for
\code{as.SummarizedExperiment}.

a \code{\link[Biobase]{ExpressionSet}}

a \code{\link[SummarizedExperiment]{SummarizedExperiment}}
}
\description{
An entire \code{FacileDataSet} or a subset of it can be converted into
bioconductor-standard assay containers, like a \code{SummarizedExperiment},
\code{DGEList}, or \code{ExpressionSet} "at any time" using various \code{as.XXX} functions,
like \code{as.DGEList(...)}.
}
\details{
We use the term "facile object" to refer to either the entirety of a
\code{FacileDataStore} or any sample-descriptor that specifies subsets of the
data, eg. where \code{fds(x)} returns a \code{FacileDataStore}. See examples for
specifics.

Note that the order that the samples and features are materialized into the
expression container are not guaranteed.
}
\examples{
fds <- exampleFacileDataSet()

# Retrieve DGEList of gene expression for all samples
y.all <- as.DGEList(fds) # gene expression of all samples

# Retrieve data for only 3 genes
# Suppose we only wanted female samples in our DGEList
y.fem <- fds \%>\%
  filter_samples(sex == "f") \%>\%
  as.DGEList() # or `as.ExpressionSet()`
}
