% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{as.DGEList}
\alias{as.DGEList}
\title{Converts a result from `fetch_expression` into a DGEList}
\usage{
as.DGEList(x, covariates = c("IC", "TC", "BCOR"), db = attr(x, "db"),
  cov.def = db[["cov.def"]], ...)
}
\arguments{
\item{x}{\code{tbl_sql} result from a call to \code{\link{fetch_expression}}}

\item{covariates}{A \code{character} vector specifying the  additional
covariates to append to \code{out$samples}. Must be valid entries in the
\code{sample_covariate::variable} column.}

\item{db}{The \code{FacilDb} object. This is extracted from \code{x} if
we are able.}

\item{cov.def}{the path to the yaml file that defines what each type of
variable is. This is also set in and extracted from \code{db}.}
}
\value{
a \code{\link[edgeR]{DGEList}}
}
\description{
The genes and samples that populate the \code{DGEList} are specified by
\code{x}, and the caller can request addition sample information to be
appended to \code{out$samples} via specification through the
\code{covariates} argument.
}

