% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity-attribute-value.R
\name{as.EAVtable}
\alias{as.EAVtable}
\title{Convert a \code{pData} data.frame to a melted EAV table}
\usage{
as.EAVtable(
  x,
  ignore = c("dataset", "sample_id"),
  covariate_def = list(),
  na.rm = TRUE
)
}
\arguments{
\item{x}{a wide \code{pData} data.frame}

\item{covariate_def}{passed to \code{\link[=eav_metadata_create]{eav_metadata_create()}} that is used to
override default covariate definitions extracted from the columns of \code{x}}
}
\value{
a melted EAV table from \code{x}
}
\description{
Transforms a wide \code{pData} data.frame into a melted EAV table for use in
a \code{FacileDataSet}. This function will also produce the list-of-list encodings
that are generated from \code{\link[=eav_metadata_create]{eav_metadata_create()}} to do its thing as an
attribute of the returned object.
}
\details{
If you want to provide custom definitions for the covariates in the EAVtable
that are different than the ones generated in \code{\link[=eav_metadata_create]{eav_metadata_create()}}, then
provie that definition list in the \code{covariate_def} parameter.
}
