% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.FacileDataSet.R
\name{as_FacileDataSet}
\alias{as_FacileDataSet}
\title{Finalizes conversion of bioconductor containers into a FacileDataSet.}
\usage{
as_FacileDataSet(sample_covariates, assays, assay_name, assay_type,
  assay_feature_type, assay_feature_info, assay_description = assay_name,
  path, meta_file, page_size = 2^12, cache_size = 2e+05,
  chunk_rows = 5000, chunk_cols = "ncol", chunk_compression = 5,
  covariate_def = list(), metayaml = NULL, ...)
}
\arguments{
\item{sample_covariates}{(named) list of pData \code{data.frame}s that will be
inserted into the \code{FacileDataSet}.}

\item{assays}{a (named) list of assay matrices for the samples across the
datasets}

\item{assay_name}{the name to use for the assay data in \code{assays}.}

\item{assay_type}{the type of assay (ie. \code{"rnaseq"}, \code{"affymetrix"}, etc.)}

\item{assay_feature_type}{the name of the "feature space" over the rows
of this assay, something like \code{"entrez"}, \code{"ensgid"}, \code{"enstid"}, etc.}

\item{assay_feature_info}{the \code{feature_info} \code{data.frame} for the features
in the rows of \code{assays}}

\item{path}{the path to the folder that will contain the facile contents}

\item{meta_file}{the path on disk where the \code{meta.yaml} file exists for
this \code{FacileDataSet}. Reference the help in \code{\link[=FacileDataSet]{FacileDataSet()}} for a more
complete description fo what is expected in the \code{meta.yaml} file.}

\item{page_size}{parameter to tweak SQLite}

\item{cache_size}{parameter to tweak SQLite}
}
\description{
This is the final call that the various
\code{as.FacileDataSet.(SummarizedExperiment)} functions delegate to in order to
actually create the \code{FacileDataSet} on disk. It is used to put all of the
minimal "bits" for a valid \code{FacileDataSet} into place, and is only meant to
initialize it with data from one assay type.
}
\details{
To add more assays to the \code{FacileDataSet}, use the \code{\link[=addFacileAssaySet]{addFacileAssaySet()}}
function.

\strong{This function is intentionally not exported}, however a savvy user may
find themselves calling it to fill out a complete FacileDataSet after it has
been initially constructed via a call to \code{as.FacileDataSet(stuff, ...)}
not exported
}
