% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay-data.R
\name{assay_feature_type}
\alias{assay_feature_type}
\title{Returns the feature_type for a given assay}
\usage{
assay_feature_type(x, assay_name)
}
\arguments{
\item{x}{\code{FacileDataSet}}

\item{assay_name}{the name of the assay}
}
\description{
The elements of the rows for a given assay all correspond to a particular
feature space (ie. feature_type='entrez')
}
