% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay-data.R
\name{assay_info_over_samples}
\alias{assay_info_over_samples}
\title{Identify the number of each assay run across specific samples}
\usage{
assay_info_over_samples(x, samples)
}
\arguments{
\item{x}{FacileDataSet}

\item{samples}{sample descriptor}

\item{with_count}{return the number of samples in \code{samples} that are
assayed over each assay as a column in \code{return}}
}
\value{
rows from assay_info_tbl that correspond to the assays defined
  over the given samples. If no assays are defined over these samples,
  you're going to get an empty tibble.
}
\description{
Identify the number of each assay run across specific samples
}
