% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{assay_name}
\alias{assay_name}
\title{What was the assay_name this result was run on?}
\usage{
assay_name(x, ...)
}
\arguments{
\item{x}{the facile object}
}
\description{
What was the assay_name this result was run on?
}
