% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay-data.R
\name{assay_sample_info}
\alias{assay_sample_info}
\title{Utility functions to get row and column indices of rnaseq hdf5 files.}
\usage{
assay_sample_info(x, assay_name, samples = NULL)
}
\description{
Utility functions to get row and column indices of rnaseq hdf5 files.
}
