% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay-data.R
\name{assay_summary}
\alias{assay_summary}
\title{Identify the number of each assay run across specific samples.}
\usage{
assay_summary(x, ...)
}
\arguments{
\item{x}{FacileDataSet}

\item{samples}{sample descriptor}

\item{with_count}{return the number of samples in \code{samples} that are
assayed over each assay as a column in \code{return}}
}
\value{
rows from assay_info_tbl that correspond to the assays defined
over the given samples. If no assays are defined over these samples,
you're going to get an empty tibble.
}
\description{
The default assay is listed first, the rest of the order is undetermined.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A tibble: 2 × 3
assay    ndatasets nsamples
<chr>        <int>    <int>
scrnaseq         3       26
snrnaseq         3       20
}\if{html}{\out{</div>}}
}
