% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{assay_units}
\alias{assay_units}
\title{Units of measure in an assay}
\usage{
assay_units(x, assay_name, normalized = FALSE, abbreviate = FALSE, ...)
}
\value{
string
}
\description{
Units of measure in an assay
}
