% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assemble_example_dataset.R
\name{assemble_example_dataset}
\alias{assemble_example_dataset}
\title{Assembles an example facile dataset to play with}
\usage{
assemble_example_dataset(
  directory = tempdir(),
  name = "ExampleRnaFacileDataSet"
)
}
\arguments{
\item{directory}{The name of the parent directory to hold the dataset}

\item{name}{A subdirectory within \code{directory} will be created using this
name.}
}
\value{
The FacileDataSet object itself.
}
\description{
This combines the airway and parathyroidSE RNA-seq datasets into a single
FacileDataSet.
}
\details{
The code here is extracted from the \code{FacileDataSet-assembly} vignette. Please
read that for some of the why's and how's of the decisions made here when
assembling datasets.
}
\examples{
\dontest{
afds <- assemble_example_dataset()
}
}
