% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{assert_sample_subset}
\alias{assert_sample_subset}
\title{Check to see that samples are referenced correctly}
\usage{
assert_sample_subset(x)
}
\description{
Samples have compound keys: dataset,sample_id. If we want to index into
them, we can either:
}
\details{
1. pass a data.frame around with dataset and sample_id columns
  2. pass a "loaded up" tbl_sqlite" over the sample_covariate table which
     has your filters of interest set
}

