% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{assert_sample_subset}
\alias{assert_sample_subset}
\alias{check_sample_subset}
\alias{test_sample_subset}
\alias{assert_facet_descriptor}
\alias{is_facet_descriptor}
\alias{assert_assay_feature_descriptor}
\alias{is_assay_feature_descriptor}
\alias{assert_expression_result}
\alias{is_expression_result}
\alias{assert_sample_statistics}
\alias{is_sample_statistics}
\alias{assert_sample_covariates}
\alias{is_sample_covariates}
\alias{assert_columns}
\alias{has_columns}
\alias{assert_covariate_definitions}
\alias{is_covariate_definitions}
\title{Check to see that samples are referenced correctly}
\usage{
assert_sample_subset(x, fds = NULL, ..., .var.name = vname(x),
  add = NULL)

check_sample_subset(x, fds = NULL, ...)

test_sample_subset(x, fds = NULL, ...)

assert_facet_descriptor(x)

is_facet_descriptor(x)

assert_assay_feature_descriptor(x, .fds = NULL)

is_assay_feature_descriptor(x, .fds = NULL)

assert_expression_result(x)

is_expression_result(x)

assert_sample_statistics(x)

is_sample_statistics(x)

assert_sample_covariates(x)

is_sample_covariates(x)

assert_columns(x, req.cols)

has_columns(x, req.cols)

assert_covariate_definitions(x)

is_covariate_definitions(x)
}
\description{
Samples have compound keys: dataset,sample_id. If we want to index into
them, we can either:
}
\details{
1. pass a data.frame around with dataset and sample_id columns
  2. pass a "loaded up" tbl_sqlite" over the sample_covariate table which
     has your filters of interest set
}
\section{assay_feature_descriptor}{

If .fds is provided, it must be a \code{FaclieDataSet} and these functions
will check to ensure that the \code{x[['assay']]} is a valid assay element
in \code{.fds}
}

