% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{bind_pdata_rows}
\alias{bind_pdata_rows}
\title{bind_rows a list of data.frames that might have Surv columns}
\usage{
bind_pdata_rows(df_list)
}
\arguments{
\item{df_list}{list of data.frames}
}
\value{
data.frame
}
\description{
Surv columns currently break bind_rows, so we coerce to character, bind_rows,
and coerce back.
}
