% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-raw-data-assembly-helpers.R
\name{build_assay_lists_load}
\alias{build_assay_lists_load}
\title{Loads an assay list object}
\usage{
build_assay_lists_load(assay_name)
}
\arguments{
\item{assay_name}{One of the (case insensitive) options in
\code{available_assays()$assay_name}}
}
\description{
Loads an assay list object
}
