% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-raw-data-assembly-helpers.R
\name{build_available_assays}
\alias{build_available_assays}
\title{Description of available assays in this packge}
\usage{
build_available_assays()
}
\value{
a tibble of assay information
}
\description{
Description of available assays in this packge
}
