% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-raw-data-assembly-helpers.R
\name{build_pb_split}
\alias{build_pb_split}
\title{Splits a SingleCellExperiment in a list of DGELists, split by the levels
one of the colData columns}
\usage{
build_pb_split(x, split.by = "cond")
}
\arguments{
\item{x}{A SingleCellExperiment or SummarizedExperiment}

\item{split.by}{a column name in \code{colData(x)} that has categorical covariate
(factor or character)}
}
\description{
This function is called from the dataset-wrangling.Rmd vignette, and
relies on the SingleCellExperiment package (which is in Suggests, not
Imports)
}
