% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_categorical}
\alias{check_categorical}
\alias{assert_categorical}
\alias{test_categorical}
\title{Check to see if a vector is categorical (character or string)}
\usage{
check_categorical(x, ...)

assert_categorical(x, ..., .var.name = vname(x), add = NULL)

test_categorical(x, ...)
}
\arguments{
\item{x}{a vector of things}

\item{...}{dots}
}
\description{
Check to see if a vector is categorical (character or string)
}
