% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_facile_data_set}
\alias{check_facile_data_set}
\alias{assert_facile_data_set}
\alias{test_facile_data_set}
\title{Check if argument is a FacileDataSet}
\usage{
check_facile_data_set(x, ...)

assert_facile_data_set(x, ..., .var.name = vname(x), add = NULL)

test_facile_data_set(x, ...)
}
\arguments{
\item{x}{The object to check.}

\item{...}{to be determined later}

\item{.var.name}{Name of the checked object to print in assertions. Defaults
to the heuristic implemented in [checkmate::vname()].}

\item{add}{An [checkmate::AssertCollection()] object. Default is `NULL`.}
}
\description{
Check if argument is a FacileDataSet
}
