% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_facile_data_store}
\alias{check_facile_data_store}
\alias{assert_facile_data_store}
\alias{test_facile_data_store}
\title{Check if argument is a FacileDataStore}
\usage{
check_facile_data_store(x, ...)

assert_facile_data_store(x, ..., .var.name = vname(x), add = NULL)

test_facile_data_store(x, ...)
}
\arguments{
\item{x}{The object to check.}

\item{...}{to be determined later}

\item{.var.name}{Name of the checked object to print in assertions. Defaults
to the heuristic implemented in \code{\link[checkmate:vname]{checkmate::vname()}}.}

\item{add}{An \code{\link[checkmate:AssertCollection]{checkmate::AssertCollection()}} object. Default is \code{NULL}.}
}
\description{
Check if argument is a FacileDataStore
}
