% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{conform_data_frame}
\alias{conform_data_frame}
\title{Arranges the columns of one data.frame to another}
\usage{
conform_data_frame(x, to)
}
\arguments{
\item{x}{a \code{data.frame} that needs to be checked and conformed}

\item{to}{the prototype \code{data.frame} that \code{x} needs to be aligned
against.}
}
\value{
the \code{tibble} version of \code{x} that is arranged to look
  like \code{to}.
}
\description{
This function is primarily used to add data to the FacileDataSet's SQLite
database. \code{x} is new data to add, and \code{to} is the a table of
the form that is expected in the database. We check that the columns of
\code{x} are a superset of columns in \code{x} and the matching columns are
all of the same class.
}
