% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FacileDataSet.R
\name{covariate_definitions}
\alias{covariate_definitions}
\title{Get description of sample metadata columns}
\usage{
covariate_definitions(x, as.list = TRUE)
}
\arguments{
\item{x}{FacileDataSet}

\item{as.list}{single logical, return tibble or list}
}
\description{
Get description of sample metadata columns
}
