% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-covariates.R
\name{covariate_meta_info}
\alias{covariate_meta_info}
\title{Retrieve the meta information about a covariate}
\usage{
covariate_meta_info(covariate, .fds, covdefs = covariate_definitions(.fds))
}
\arguments{
\item{covariate}{the name of the covariate}

\item{.fds}{the \code{FacileDataSet}}

\item{covdefs}{The \code{covariate_definitions(.fds)} list}
}
\value{
a list of covariate information with the following elements:
  \code{$name}, \code{$type}, \code{$class}, \code{$description},
  \code{$label}, \code{$is.factor}, (and maybe \code{$levels})
}
\description{
Retrieve the meta information about a covariate
}
