% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-covariates.R
\name{covariate_summary.FacileDataSet}
\alias{covariate_summary.FacileDataSet}
\title{Provides a summary table of the (sample) covariates in a FacileDataStore}
\usage{
covariate_summary.FacileDataSet(x, samples = active_samples(x),
  custom_key = Sys.getenv("USER"), covariate_type = c("samples"),
  with_source = FALSE, ...)
}
\arguments{
\item{x}{a \code{FacileDataSet} connection}

\item{samples}{a samples descriptor, used to narrow the universe of samples
to get sample covariate information from.}

\item{...}{extras}
}
\value{
a tibble of summary sample covariate information with the following
  columns:
  * `variable`: name of the variable
  * `class`: class of variable (real, categorical)
  * `nsamples`: the number of samples that have this variable defined
}
\description{
Provides a summary table of the (sample) covariates in a FacileDataStore
}
\seealso{
Other FacileInterface: \code{\link{facet_frame.FacileDataSet}},
  \code{\link{organism}}, \code{\link{sample_covariates}}
}
\concept{FacileInterface}
