% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.FacileDataSet.R
\name{covdef_for_column}
\alias{covdef_for_column}
\title{Generate a generic covariate definition for a column in a data.frame}
\usage{
covdef_for_column(x, column)
}
\arguments{
\item{x}{a `data.frame`}

\item{column}{the name of the column in the `x` to parse.}
}
\value{
a generic list-of-list defintion for `x[[column]]`
}
\description{
Generate a generic covariate definition for a column in a data.frame
}
