% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.FacileDataSet.R
\name{covdefs_from_df}
\alias{covdefs_from_df}
\title{Create a facile covariate definition file from a sample covariate data.frame}
\usage{
covdefs_from_df(x, covariate_def = list())
}
\arguments{
\item{x}{a `pData` `data.frame`}

\item{covariate_def}{a named list of covariate definitions. The names of
this list are the names the covariates will be called in the target
`FacileDataSet`. The values of the list are:
* `varname`: a `character()` of the column name(s) in `x` that this
   sample covariate was derived from. If more than one column is to be used
   for the facile covariate conversion (eg. if we are encoding survival),
   then provide a `length() > 1` character vector with the names of the
   columns in `x` that were used for the encoding. If this were encoding
   survival this might be `c("time", "event") columns, in that order.
* label: a human readable label to use for this covariate in user facing
   scenarios in the facileverse.
* `class`: the "facile class" of the covariate. This can either be
  `categorical`, `real`, or `right_censored` (for survival).
* `levels`: (optional) if you want a `categorical` to be treated as a
  factor if it isn't already encoded as such in the `pData` itself, or if
  you want to rearrange the factor levels.
* `type`: (optinal) this is used a a "grouping" level, particularly in
  the FacileExplorer. Not including this won't matter.
  TODO: talk about covariate groupings in
  `FacileExplorer::categoricalCovariateSelect`}
}
\value{
a list-of-lists that encodes the `sample_covariate` section of the
  `meta.yaml` file for a `FacileDataSet`.
}
\description{
Sample covariates (aka `pData`) need are encoded in an
entity-attrivute-value table. Metadata about these covariates are stored
in a `meta.yaml` file in the FacileDataSet directory. This function generates
the list-of-list structure to represent the `sample_covariates` section of
the `meta.yaml` file.
}
\section{Survival}{

Survival data is encoded by two columns. One column to indicate the
"time to event" (tte), and a second to indicate whether or not the denoted
tte is an "event" (1) or "censored" (0). If there are such data in `x`, it
must be in a (`tte_OS`, `event_OS`) pair of columns for "ordinary survival"
or a (`tte_PFS`, `event_PFS`) for progression free survival.
}

\examples{
# covariate_def definition to take tte_OS and tte_event columns and turn
# into a facile "OS" right_censored survival covariate
cc <- list(
  OS=list(
    varname=c("tte_OS", "tte_event"),
    label="Overall Survival",
    class="right_censored",
    type="clinical",
    description="Overall survival in days"))
}
