% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay-data.R, R/expression.R
\name{cpm.tbl_sqlite}
\alias{cpm.tbl_sqlite}
\alias{cpm.tbl_df}
\alias{cpm.tbl_sqlite}
\alias{cpm.tbl_df}
\title{Calculated counts per million from a FacileDataSet}
\usage{
\method{cpm}{tbl_sqlite}(x, lib.size = NULL, log = FALSE, prior.count = 5,
  feature_ids = NULL, as.matrix = FALSE, .fds = fds(x), ...)

\method{cpm}{tbl_df}(x, lib.size = NULL, log = FALSE, prior.count = 5,
  feature_ids = NULL, sample.stats = NULL, as.matrix = as.matrix,
  .fds = fds(x), ...)

\method{cpm}{tbl_sqlite}(x, lib.size = NULL, log = FALSE, prior.count = 5,
  feature_ids = NULL, as.matrix = FALSE, .fds = fds(x), ...)

\method{cpm}{tbl_df}(x, lib.size = NULL, log = FALSE, prior.count = 5,
  feature_ids = NULL, sample.stats = NULL, as.matrix = as.matrix,
  .fds = fds(x), ...)
}
\arguments{
\item{x}{an expression-like facile result}

\item{lib.size}{ignored for now, this is fetched from the
\code{FacileDataSet}}

\item{log}{log the result?}

\item{prior.count}{prior.count to add to observed counts}

\item{.fds}{A \code{FacileDataSet} object}

\item{x}{an expression-like facile result}

\item{lib.size}{ignored for now, this is fetched from the
\code{FacileDataSet}}

\item{log}{log the result?}

\item{prior.count}{prior.count to add to observed counts}

\item{.fds}{A \code{FacileDataSet} object}
}
\value{
a modified expression-like result with a \code{cpm} column.

a modified expression-like result with a \code{cpm} column.
}
\description{
Calculated counts per million from a FacileDataSet

Calculated counts per million from a FacileDataSet
}
