% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression-sqlite.R
\name{cpmdb.tbl_sqlite}
\alias{cpmdb.tbl_df}
\alias{cpmdb.tbl_sqlite}
\title{Calculated counts per million from a FacileDataSet}
\usage{
\method{cpmdb}{tbl_sqlite}(x, lib.size = NULL, log = FALSE,
  prior.count = 5, .fds = fds(x), ...)

\method{cpmdb}{tbl_df}(x, lib.size = NULL, log = FALSE, prior.count = 5,
  sample.stats = NULL, .fds = fds(x), ...)
}
\arguments{
\item{x}{an expression-like facile result}

\item{lib.size}{ignored for now, this is fetched from the
\code{FacileDataSet}}

\item{log}{log the result?}

\item{prior.count}{prior.count to add to observed counts}

\item{.fds}{A \code{FacileDataSet} object}
}
\value{
a modified expression-like result with a \code{cpm} column.
}
\description{
Calculated counts per million from a FacileDataSet
}

