% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay-data.R
\name{create_assay_feature_descriptor}
\alias{create_assay_feature_descriptor}
\title{Creates a feature descriptor for interactive ease}
\usage{
create_assay_feature_descriptor(x, features = NULL, assay_name = NULL)
}
\arguments{
\item{x}{FacileDataSet}

\item{features}{a character string of fearture ids (requires assay_name)
or a data.frame with feature_id column.}

\item{assay_name}{the assay to get the featurespace from. if this is provided,
it will trump an already existing assay_name column in \code{features}}
}
\value{
a feature descriptor with feature_id and assay_name, which can be
  used to absolutely find features
}
\description{
Creates a data.frame of features and assays they come from
}
