% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc-assay-containers.R
\name{create_metayaml}
\alias{create_metayaml}
\title{Creates a shell of a yaml file for a FacileDataSet}
\usage{
create_metayaml(name = "unspecified", organism = "unspecified",
  datasets = list(), sample_covariates = list(),
  default_assay = "unspecified")
}
\description{
Creates a shell of a yaml file for a FacileDataSet
}
