% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5.R
\name{createExpressionHDF5}
\alias{createExpressionHDF5}
\title{Creates an HDF5 enabled FacileDataSet from a completely sqlite-based one.}
\usage{
createExpressionHDF5(x, drop.expr.table = TRUE)
}
\description{
This is a convenience function that will be axed eventually.
}
