% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-db.R
\name{createExpressionTables}
\alias{createExpressionTables}
\title{Creates the table that holds gene counts and meta information about data}
\usage{
createExpressionTables(db)
}
\description{
Creates the table that holds gene counts and meta information about data
}
