% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construction.R
\name{createFacileDataRepository}
\alias{createFacileDataRepository}
\title{Create a new FacileDataRepository}
\usage{
createFacileDataRepository(path, datas, covariate_definition, data.cov.map,
  page_size = 2^12, cache_size = 2e+05)
}
\arguments{
\item{path}{the path to the directory you want to create to put the
FacileDataRepository into.}

\item{datas}{a list of data objects. Currently objects in the list can only
either be ExpressoinSets or SummarizedExperiments.}

\item{covariate_definition}{the YAML file that maps variables that land into
the sample_covariate table into the R-alike things}

\item{data.cov.map}{a YAML file that maps pData elements in the datas list
to variable/value pairs into the smaple_covariate table.}
}
\value{
the path to the FacileDataRepository directory
}
\description{
Create a new FacileDataRepository
}
