% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FacileDataSet.R
\name{dbfn}
\alias{dbfn}
\title{Get location of the FacileDataSet database}
\usage{
dbfn(x, mustWork = TRUE)
}
\arguments{
\item{x}{FacileDataSet}

\item{mustWork}{single logical}
}
\value{
single character, path to database
}
\description{
Get location of the FacileDataSet database
}
