% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{dbGetQueries}
\alias{dbGetQueries}
\title{Execute multiple queries against a database}
\usage{
dbGetQueries(con, sql)
}
\arguments{
\item{con}{database handle}

\item{sql}{list of charvecs (SQL statements)}
}
\description{
Execute multiple queries against a database
}
