% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{decode_right_censored}
\alias{decode_right_censored}
\title{Convert +/- encoding of time to event to time + event.}
\usage{
decode_right_censored(x, suffix = NULL)
}
\arguments{
\item{x}{the time to event}
}
\value{
two column data.frame
}
\description{
Time to event (and event labels) are encoded as a single real value in the
clinical data. The absoulte value of the value is the time to the event, and
a negative value indicates "event" and a positive value is right censored.
}

