% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_transcribed_info_from_ensembl_gtf.R
\name{.level_biotypes}
\alias{.level_biotypes}
\title{Utility function to "factorize" biotypes into an order we care about.}
\usage{
.level_biotypes(x)
}
\arguments{
\item{x}{a \code{character} vector of biotypes}
}
\value{
a factor version of \code{x}, with \code{levels(x)} in approximately the order
we care about.
}
\description{
ENSEMBL GTFs provide biotype information for genes/transcripts. These are
things like "3prime_overlapping_ncRNA", "antisense", ..., "protein_coding",
etc. This function turns the "biotype"-vector \code{x} into a factor with levels
in (roughly) the order we care to "unique"-ify these levels. Ie. if a gene
has a "protein_coding" annotation, we will care to keep that one over one
of its annotations which categorize it as a "processed_transcript"
}
