% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.FacileDataSet.R
\name{.prepare_sample_covariates}
\alias{.prepare_sample_covariates}
\title{Helper function when creating or appending to a faciledatsets}
\usage{
.prepare_sample_covariates(
  x,
  covariate_def = NULL,
  fds = NULL,
  parents = NULL,
  ...
)
}
\arguments{
\item{x}{a named list of assay containers (names are the \code{dataset} values)}

\item{covariate_def}{precanned covariate definitions}

\item{fds}{(optional) a \code{FacileDataSet} object that you are adding these
sample level covariates to.}

\item{parents}{not used now, but a two-column data.frame that encodes which
sample is a parent of another}
}
\description{
Preapres the EAV table and stuff for the sample-level covariates (pData) of
the assay objects in the named list \code{x}
}
\details{
TODO: Handle and optional parameter that provides new or existing mappings
mappings of covariate names to the attribute-value definitions that
eventually get serialized to the meta.yaml file
}
