% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.FacileDataSet.R
\name{ds_annot}
\alias{ds_annot}
\title{Bioc-container specific data set annotation extraction functions}
\usage{
ds_annot(x, validate = FALSE, ...)
}
\description{
Takes dataset-level annotion as stored by each type. DGEList has
no such slot, unfortunately, and thus gets the default. SE has a
metadata slot and can provide url and description. eSet just has
a character annotation and can provide a description.
}
