% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.FacileDataSet.R
\name{create_eav_metadata}
\alias{create_eav_metadata}
\alias{validate_covariate_def_list}
\title{Create a facile covariate definition file from a sample \code{pData} data.frame}
\usage{
create_eav_metadata(x, covariate_def = list())

validate_covariate_def_list(x, pdata)
}
\arguments{
\item{x}{a \code{pData} \code{data.frame}}

\item{covariate_def}{a named list of covariate definitions. The names of
this list are the names the covariates will be called in the target
\code{FacileDataSet}. The values of the list are:
\itemize{
\item \code{varname}: a \code{character()} of the column name(s) in \code{x} that this
sample covariate was derived from. If more than one column is to be used
for the facile covariate conversion (eg. if we are encoding survival),
then provide a \code{length() > 1} character vector with the names of the
columns in \code{x} that were used for the encoding. If this were encoding
survival this might be `c("time", "event") columns, in that order.
\item label: a human readable label to use for this covariate in user facing
scenarios in the facileverse.
\item \code{class}: the "facile class" of the covariate. This can either be
\code{categorical}, \code{real}, or \code{right_censored} (for survival).
\item \code{levels}: (optional) if you want a \code{categorical} to be treated as a
factor if it isn't already encoded as such in the \code{pData} itself, or if
you want to rearrange the factor levels.
\item \code{type}: (optinal) this is used a a "grouping" level, particularly in
the FacileExplorer. Not including this won't matter.
TODO: talk about covariate groupings in
\code{FacileExplorer::categoricalCovariateSelect}
}}
}
\value{
a list-of-lists that encodes the \code{sample_covariate} section of the
\code{meta.yaml} file for a \code{FacileDataSet}.
}
\description{
Sample covariates (aka \code{pData}) are encoded in an
\href{https://en.wikipedia.org/wiki/Entity-attribute-value_model}{entity-attribute-value (EAV) table}.
Metadata about these covariates are stored in a \code{meta.yaml} file in the
\code{FacileDataSet} directory which enables the \code{FacileDataSet} to cast the value
stored in the EAV table to its native R type. This function generates the
list-of-list structure to represent the \code{sample_covariates} section of the
\code{meta.yaml} file.
}
\details{
For simple \code{pData} covariates, each column is treated independantly from the
rest. There are some types of covariates which require multiple columns for
proper encoding, such as encoding survival information (see the \strong{Survival}
section below). In these cases, the caller needs to provide an entry in the
\code{covariate_def} list that describes which \code{pData} columns (\code{varname}) goes
into the single facile covariate value. Please reference the
\strong{Defining EAV Encodings} section for more information.

keeping track of the "time to event" in one column, and a separae column to
indicate whether or not the event was a "death" or a censored. Still, these
data are stored in the single "value" column of the FacileDataSet's internal
entity-attribute-value (\code{sample_covariate}) table. In order to
encode these types of columns correctly, we need to provide more information
via the \code{covariate_def} parameter of this function.
}
\section{Survival}{

Survival data is encoded by two columns. One column to indicate the
"time to event" and a second to indicate whether or not the denoted
tte is an "event" (1) or "censored" (0). The pair of columns will be encoded
into the \code{FacileDataSet}'s EAV table as a single (numeric) value. The
absolute value of the numeric indicates the "time to event" and the sign of
the value indicates its censoring status (If there are such data in \code{x}, it
must be in a (\code{tte_OS}, \code{event_OS}) pair of columns for "ordinary survival"
or a (\code{tte_PFS}, \code{event_PFS}) for progression free survival. Please see the
"Defining EAV Encodings" sections for more details.
}

\section{Defining EAV Encodings}{

}

\examples{
# covariate_def definition to take tte_OS and tte_event columns and turn
# into a facile "OS" right_censored survival covariate
cc <- list(
  OS=list(
    varname=c("tte_OS", "tte_event"),
    label="Overall Survival",
    class="right_censored",
    type="clinical",
    description="Overall survival in days"))
}
