% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity-attribute-value.R
\name{eavdef_for_column}
\alias{eavdef_for_column}
\title{Generate entity-attribute-value definition for a column in a data.frame}
\usage{
eavdef_for_column(x, column)
}
\arguments{
\item{x}{a \code{data.frame}}

\item{column}{the name of the column in the \code{x} to parse.}
}
\value{
a generic list-of-list defintion for \code{x[[column]]}

[[column]: R:[column
}
\description{
Creates the miniimal list-definition for a single column in a \code{pData}
\code{data.frame}. This function is not exported on purpose.
}
