% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{executeSQL}
\alias{executeSQL}
\title{Execute multiple queries against a database}
\usage{
executeSQL(con, sql)
}
\arguments{
\item{con}{database handle}

\item{sql}{list of charvecs (SQL statements)}
}
\description{
Execute multiple queries against a database
}
