% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc-assay-containers.R
\name{as.DGEList}
\alias{as.DGEList}
\alias{as.DGEList.matrix}
\alias{as.DGEList.data.frame}
\alias{as.DGEList.tbl_sql}
\alias{as.ExpressionSet}
\alias{as.ExpressionSet.data.frame}
\alias{as.ExpressionSet.FacileDataSet}
\alias{as.SummarizedExperiment}
\alias{as.SummarizedExperiment.data.frame}
\alias{as.SummarizedExperiment.FacileDataSet}
\title{Converts a facile result into a DGEList or ExpressionSet, or ...}
\usage{
as.DGEList(x, ...)

\method{as.DGEList}{matrix}(x, covariates = TRUE, feature_ids = NULL,
  assay_name = default_assay(.fds), .fds = fds(x),
  custom_key = Sys.getenv("USER"), ...)

\method{as.DGEList}{data.frame}(x, covariates = TRUE, feature_ids = NULL,
  assay_name = default_assay(.fds), .fds = fds(x),
  custom_key = Sys.getenv("USER"), ...)

\method{as.DGEList}{tbl_sql}(x, covariates = TRUE, feature_ids = NULL,
  assay_name = default_assay(.fds), .fds = fds(x),
  custom_key = Sys.getenv("USER"), ...)

as.ExpressionSet(x, ...)

\method{as.ExpressionSet}{data.frame}(x, covariates = TRUE,
  feature_ids = NULL, assay_name = default_assay(.fds), .fds = fds(x),
  custom_key = Sys.getenv("USER"), ...)

\method{as.ExpressionSet}{FacileDataSet}(x, covariates = TRUE,
  feature_ids = NULL, assay_name = default_assay(.fds), .fds = fds(x),
  custom_key = Sys.getenv("USER"), ...)

as.SummarizedExperiment(x, ...)

\method{as.SummarizedExperiment}{data.frame}(x, covariates = TRUE,
  feature_ids = NULL, assay_name = default_assay(.fds), .fds = fds(x),
  custom_key = Sys.getenv("USER"), ...)

\method{as.ExpressionSet}{FacileDataSet}(x, covariates = TRUE,
  feature_ids = NULL, assay_name = default_assay(.fds), .fds = fds(x),
  custom_key = Sys.getenv("USER"), ...)
}
\arguments{
\item{x}{a facile expression-like result}

\item{covariates}{The covariates the user wants to add to the $samples of
the DGEList. This can take the following forms:
\enumerate{
  \item{TRUE}{
    All covariates are retrieved from the \code{FacileDataSet}
  }
  \item{FALSE}{
    TODO: Better handle FALSE
  }
  \item{character}{
    A vector of covariate names to fetch from the \code{FacileDataSet}
  }
  \item{data.frame}{
    A table that looks like a subset of the sample_covariate table
  }
  \item{NULL}{
    If \code{NULL}, no covariates are retrieved
  }
}}

\item{feature_ids}{the features to get expression for (if not specified
in \code{x} descriptor)}

\item{.fds}{The \code{FacileDataSet} that \code{x} was retrieved from.}

\item{custom_key}{the custom key to use to fetch custom annotations from
\code{.fds}}

\item{assay}{the \code{assayDataElement} to use for the expression data
if \code{x} is an \code{ExpressionSet}.}
}
\value{
a \code{\link[edgeR]{DGEList}}

a \code{\link[Biobase]{ExpressionSet}}

a \code{\link[SummarizedExperiment]{SummarizedExperiment}}
}
\description{
The genes and samples that populate the \code{DGEList} are specified by
\code{x}, and the caller can request addition sample information to be
appended to \code{out$samples} via specification through the
\code{covariates} argument.
}
