% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_transcribed_info_from_ensembl_gtf.R
\name{extract_transcribed_info_from_ensembl_gtf}
\alias{extract_transcribed_info_from_ensembl_gtf}
\title{Extract gene- and transcript-level information from an ENSEMBL gtf.}
\usage{
extract_transcribed_info_from_ensembl_gtf(fn)
}
\arguments{
\item{fn}{the path to the ENSEMBL (or GENCODE) GTF}
}
\value{
a list of tibbles with \code{$transcript_info} and \code{$gene_info} elements
}
\description{
This was written for release_28 annotations. This is noted because some
column names seemsed to have changed, ie. "gene_type" instead of
"gene_biotype", etc. Let's see how consistent this is!
}
