% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FacileDataSet.R
\name{facet_frame}
\alias{facet_frame}
\alias{facet_frame.default}
\alias{facet_frame.FacileDataStore}
\title{Retrieves grouping table for samples within a FacileDataSet.}
\usage{
facet_frame(x, name = "default", ...)

\method{facet_frame}{default}(x, name = "default", ...)

\method{facet_frame}{FacileDataStore}(x, name = "default", ...)
}
\arguments{
\item{x}{A `FacileDataStore`}

\item{name}{The specific facet (grouping) definition to return. Note that
this parameter isn't yet used. Only one facet table was originally
defined for each FacileDataSet, but we want to enable different facet
definitions to be used in the future.}
}
\value{
A `tibble` that defines the `dataset,sample_id` tuples that belong
  to each "facet" (group).
}
\description{
It is natural to define subgroups of samples within larger datasets.
This function returns grouping definitions (which we call "facets") for
a `FacileDataStore`.
}
\seealso{
Other API: \code{\link{fetch_assay_data}},
  \code{\link{fetch_assay_score}},
  \code{\link{fetch_custom_sample_covariates}},
  \code{\link{fetch_sample_covariates}},
  \code{\link{fetch_sample_statistics}},
  \code{\link{fetch_samples}},
  \code{\link{filter_features}},
  \code{\link{filter_samples}}, \code{\link{organism}},
  \code{\link{samples}}
}
\concept{API}
