% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{FacileDataSet-package}
\alias{FacileDataSet}
\alias{FacileDataSet-package}
\title{More description goes here about the FacileDataSet package.}
\description{
More description goes here about the FacileDataSet package.
}
\author{
\strong{Maintainer}: Steve Lianoglou \email{slianoglou@gmail.com}

Other contributors:
\itemize{
  \item Vincent Rouilly \email{rouilly.vincent@gene.com} [contributor]
  \item Genentech [copyright holder, funder]
}

}
