% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{FacileDb}
\alias{FacileDb}
\title{Setup the SQLite connection the TCGA database}
\usage{
FacileDb(db.path = getOption("fatezo.dbpath", NULL))
}
\arguments{
\item{db.path}{The path to the clinX database.}
}
\value{
A \code{dplyr::src_sqlite} connection to the database.
}
\description{
Setup the SQLite connection the TCGA database
}

