% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.FacileDataSet.R
\name{fdata}
\alias{fdata}
\title{BioC-container specific fData extraction functions}
\usage{
fdata(x, validate = FALSE, ...)
}
\arguments{
\item{x}{SummarizedExperiment, ExpressionSet or DGEList}

\item{validate}{single logical, check results}

\item{...}{additional args (ignored for now)}
}
\description{
BioC-container specific fData extraction functions
}
