% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{fds}
\alias{fds}
\alias{fds.FacileDataStore}
\alias{fds.default}
\alias{fds<-}
\alias{fds<-.tbl}
\alias{fds<-.data.frame}
\alias{set_fds}
\title{Get or set the FacileDataStore for an object}
\usage{
fds(x, ...)

\method{fds}{FacileDataStore}(x)

\method{fds}{default}(x, ...)

fds(x) <- value

\method{fds}{tbl}(x) <- value

\method{fds}{data.frame}(x) <- value

set_fds(x, value)
}
\arguments{
\item{x}{the object}

\item{value}{The \code{FacileDataStore} object}
}
\description{
FacileDataStores are passed along with most every object generated from
functions in the facilebio universe. This makes it convenient to dig back
into a large genomics objects to retrieve data from "slim" results, like
a sample covariate data.frame.
}
