% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fds-creation.R
\name{fds_add_assay_data}
\alias{fds_add_assay_data}
\title{Add an assay matrix to a pre-registered assay in an FDS.}
\usage{
fds_add_assay_data(
  x,
  assay_matrix,
  assay_name,
  dataset_name,
  ...,
  add_samples = TRUE,
  missing_value = 0L,
  chunk_rows = 5000,
  chunk_cols = "ncol",
  chunk_compression = 4
)
}
\arguments{
\item{x}{a FacileDataSet}

\item{assay_name}{the name of a pre-registered assay in \code{x}}

\item{dataset_name}{a string indicating the \code{dataset} value for this data.
Recall that dataset,sample_id (ie. dataset,colnames(assay_matrix)) are
the primary key's of the data in \code{x}}

\item{...}{stuff}

\item{add_samples}{if TRUE (default) the dataset,sample combinations are
added to the sample_info table}

\item{missing_value}{what to fill in to the internal hdf5 assay matrix for
features that are not present in \code{assay_matrix} but exist in the
\code{assay_name} feature space. Default is 0}

\item{data_matrix}{the matrix of assay_data to add. \code{rownames(data_matrix)}
must be already existing \code{feature_id}s in the registered \code{assay_name}}
}
\description{
This functions adds data to an already registered assay. The rownames() of
the assay data must already exist in the feature space that is assigned to
the assay.
}
\details{
Rows in \code{assay_matrix} that do not exist in the registered feature space
will be ignored.
}
