% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fds-creation.R
\name{fds_add_feature_space}
\alias{fds_add_feature_space}
\title{Add a new feature space to add assay data for}
\usage{
fds_add_feature_space(
  x,
  feature_info,
  description = NULL,
  feature_type = NULL,
  assays = NULL,
  ...
)
}
\arguments{
\item{x}{a faciledataset}

\item{feature_info}{a tibble of the feature universe, minimally must contain
\code{feature_id}, \code{name}, \code{meta} columns. If \code{feature_type} is not a column,
you can specify it from the \code{feature_type} parameter.}

\item{feature_type}{if provided, it will overvide the value in \code{feature_info}
If not provided (default), function will fail if \code{feature_type} is not
provided as a coloumn in \code{feature_info}}

\item{assays}{if provided, this needs to be a tibble with \code{assay_name},
\code{assay_type}, (and optional \code{description}) columns -- this will auto
register this feature space with the enumerated assays.}
}
\description{
Add a new feature space to add assay data for
}
\examples{
afds <- an_fds()
fdir <- tempfile("fds__")
afeatures <- features(afds)

xfds <- fds_initialize(fdir, "human")
fnew <- fds_add_feature_space(xfds, afeatures)
}
