% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fds-creation.R
\name{fds_add_sample_data}
\alias{fds_add_sample_data}
\title{Add (or update?) sample level metadata}
\usage{
fds_add_sample_data(
  x,
  sample_data,
  update = FALSE,
  covariate_def = NULL,
  parents = NULL,
  ...,
  validate_metadata = TRUE
)
}
\arguments{
\item{x}{a FacileDataSet}

\item{sample_data}{a tibble of data,sample_id,... columns}

\item{update}{If FALSE (default), any covariates that already exist for
samples in the database will not be updated}
}
\description{
Add (or update?) sample level metadata
}
