% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fds-creation.R
\name{fds_initialize}
\alias{fds_initialize}
\title{Initialize a new home for a FacileDataSet}
\usage{
fds_initialize(directory, species, name = NULL, ...)
}
\description{
Initialize a new home for a FacileDataSet
}
\examples{
fdir <- tempfile("fds__")
xfds <- fds_initialize(fdir, "human")
}
