% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fds-creation.R
\name{fds_modify_meta_dataset}
\alias{fds_modify_meta_dataset}
\title{Adds dataset-level metadata to a FacileDataSet}
\usage{
fds_modify_meta_dataset(x, dataset_name, ..., .nowrite = FALSE)
}
\arguments{
\item{x}{a faciledataset}

\item{dataset_name}{the name (string) of the dataset}

\item{...}{metadata variable to write to the meta.yaml file. arguments that
are strings and are prefixed with \code{dataset_} will be added. The \code{dataset_}
prefix is removed when serialized so \code{dataset_description} will be saved
as datasets.<dataset_name>.description in the yaml file}

\item{.nowrite}{even if there are changes to be made to the yaml file, we
won't reserialize it. this is used for testingn purposes so we don't
overwrite a the meta.yaml files when testing/debugging}
}
\value{
a named list of dataset metadata variables that were changed as a
result of this function call
}
\description{
Extracts all named arguments with dataset_* prefix that are strings and adds
it to the meta.yamll::datasets section for a dataset
}
