% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fds-creation.R
\name{fds_register_assay}
\alias{fds_register_assay}
\title{Register a new assay to a feature space}
\usage{
fds_register_assay(
  x,
  assay_name,
  assay_type,
  feature_type,
  ...,
  storage_mode = c("numeric", "integer")
)
}
\arguments{
\item{x}{FacileDataSet}

\item{assay_name}{the name of the new assay to register}

\item{assay_type}{the type of assay (rnaseq, lognorm, etc.)}

\item{feature_type}{the identifer of one of the featurespaces in
\code{feature_space()}}
}
\description{
Register a new assay to a feature space
}
\examples{
fds_register_assay(x, "counts", "rnaseq", "ensgid")
}
