% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay-data.R, R/features.R
\name{assay_feature_name_map}
\alias{assay_feature_name_map}
\alias{feature_name_map}
\title{Returns table of names and aliases for features.}
\usage{
assay_feature_name_map(x, assay_name)

feature_name_map(x, feature_type)
}
\arguments{
\item{x}{\code{FacileDataSet}}

\item{assay_name}{the name of assay to get the feature map for.}

\item{feature_type}{a character vector specifying the feature type}

\item{x}{\code{FacileDataSet}}
}
\value{
a tibble with \code{feature_id, name, type} columns, where type
  is "primary" or "alias"
}
\description{
Returns table of names and aliases for features.
}
