% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{feature_type}
\alias{feature_type}
\title{What was the feature_type result was run on?}
\usage{
feature_type(x, ...)
}
\arguments{
\item{x}{the facile object}
}
\description{
What was the feature_type result was run on?
}
