% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{feature_types}
\alias{feature_types}
\title{Enumerate the types of feature stored in a FacileDataSet}
\usage{
feature_types(x, with_assay_data_only = FALSE, ...)
}
\arguments{
\item{x}{A \code{FacileDataSet}}
}
\description{
Enumerate the types of feature stored in a FacileDataSet
}
