% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{features}
\alias{features}
\title{Returns a table of information about the features (from an assay, or ...)}
\usage{
features(x, ...)
}
\arguments{
\item{x}{a facile object}
}
\value{
a tibble with containing feature_id, feature_type, and whatever other
columns are appropriate given \code{x}
}
\description{
Returns a table of information about the features (from an assay, or ...)
}
