% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay-data.R
\name{fetch_assay_data}
\alias{fetch_assay_data}
\title{Fetch data from single assay of choice}
\usage{
fetch_assay_data(x, assay_name, feature_ids = NULL, samples = NULL,
  normalized = TRUE, as.matrix = FALSE, ..., .subset.threshold = 700)
}
\arguments{
\item{x}{A \code{FacileDataSet} object.}

\item{assay_name}{the name of the assay}

\item{feature_ids}{a character vector of feature_ids}

\item{samples}{a sample descriptor to specify which samples to return data
from.}

\item{normalized}{return normalize or raw data values, defaults to
\code{raw}}

\item{as.matrix}{by default, the data is returned in a long-form tbl-like
result. If set to \code{TRUE}, the data is returned as a matrix.}

\item{...}{parameters to pass to normalization methods}
}
\value{
A lazy \code{\link[dplyr]{tbl}} object with the expression
  data to be \code{\link[dplyr]{collect}}ed when \code{db} is provided,
  othwerise a \code{tbl_df} of the results.
}
\description{
Fetch data from single assay of choice
}
