% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay-data.R
\name{fetch_assay_score.FacileDataSet}
\alias{fetch_assay_score.FacileDataSet}
\title{Helper function to get sample assay data from single or aggregate features}
\usage{
\method{fetch_assay_score}{FacileDataSet}(
  x,
  features,
  samples = NULL,
  assay_name = NULL,
  as.matrix = FALSE,
  ...,
  subset.threshold = 700
)
}
\description{
Helper function to get sample assay data from single or aggregate features
}
\seealso{
Other API: 
\code{\link{fetch_custom_sample_covariates.FacileDataSet}()},
\code{\link{fetch_sample_covariates}()},
\code{\link{fetch_sample_statistics.FacileDataSet}()},
\code{\link{fetch_samples.FacileDataSet}()},
\code{\link{filter_features.FacileDataSet}()},
\code{\link{filter_samples.FacileDataSet}()},
\code{\link{organism.FacileDataSet}()},
\code{\link{samples.FacileDataSet}()}
}
\concept{API}
