% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay-data.R
\name{fetch_assay_score}
\alias{fetch_assay_score}
\title{Helper function to get sample assay data from single or aggregate features}
\usage{
fetch_assay_score(x, features, samples = NULL, assay_name = NULL,
  as.matrix = FALSE, ..., subset.threshold = 700)
}
\description{
Helper function to get sample assay data from single or aggregate features
}
\seealso{
Other API: \code{\link{fetch_assay_data}},
  \code{\link{fetch_custom_sample_covariates}},
  \code{\link{fetch_sample_covariates}},
  \code{\link{fetch_sample_statistics}},
  \code{\link{fetch_samples}},
  \code{\link{filter_features}}
}
\concept{API}
