% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-covariates.R
\name{fetch_custom_sample_covariates.FacileDataSet}
\alias{fetch_custom_sample_covariates.FacileDataSet}
\title{Fetches custom (user) annotations for a given user prefix}
\usage{
\method{fetch_custom_sample_covariates}{FacileDataSet}(x, samples = NULL,
  covariates = NULL, custom_key = Sys.getenv("USER"),
  file.prefix = "facile")
}
\arguments{
\item{samples}{the facile sample descriptor}

\item{custom_key}{The key to use for the custom annotation}

\item{fds}{The \code{FacileDataSet}}
}
\value{
covariate tbl
}
\description{
Fetches custom (user) annotations for a given user prefix
}
\seealso{
Other API: \code{\link{fetch_assay_data.FacileDataSet}},
  \code{\link{fetch_assay_score.FacileDataSet}},
  \code{\link{fetch_sample_covariates.FacileDataSet}},
  \code{\link{fetch_sample_statistics.FacileDataSet}},
  \code{\link{fetch_samples.FacileDataSet}},
  \code{\link{filter_features.FacileDataSet}},
  \code{\link{filter_samples.FacileDataSet}},
  \code{\link{organism.FacileDataSet}},
  \code{\link{samples.FacileDataSet}}
}
\concept{API}
