% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-covariates.R
\name{fetch_custom_sample_covariates}
\alias{fetch_custom_sample_covariates}
\title{Fetches custom (user) annotations for a given user prefix}
\usage{
fetch_custom_sample_covariates(x, samples = NULL, covariates = NULL,
  custom_key = Sys.getenv("USER"), file.prefix = "facile")
}
\arguments{
\item{samples}{the facile sample descriptor}

\item{custom_key}{The key to use for the custom annotation}

\item{fds}{The \code{FacileDataSet}}
}
\value{
covariate tbl
}
\description{
Fetches custom (user) annotations for a given user prefix
}

