% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{fetch_expression}
\alias{fetch_expression}
\title{Helper function creates dplyr query to get expression data of interest.}
\usage{
fetch_expression(x, samples = NULL, feature_ids = NULL,
  with_symbols = !as.matrix, as.matrix = FALSE)
}
\arguments{
\item{x}{A \code{FacileDataSet} object.}

\item{samples}{a sample descriptor to specify which samples to return data
from.}

\item{feature_ids}{character vector of ids, if \code{NULL} returns all
features}

\item{with_symbols}{adds a symbol column to the outgoing result}

\item{as.matrix}{by default, the data is returned in a long-form tbl-like
result. If set to \code{TRUE}, the data is returned as a matrix.}
}
\value{
A lazy \code{\link[dplyr]{tbl}} object with the expression
  data to be \code{\link[dplyr]{collect}}ed when \code{db} is provided,
  othwerise a \code{tbl_df} of the results.
}
\description{
Helper function creates dplyr query to get expression data of interest.
}
\section{Software Design Caveats}{


Note that if \code{db} is not provided, the result will be \code{collect}ed,
because the database connection will be closed when the function exits.

If a \code{samples} selector is provided, we currently do not make sure that
each "observation" (dataset,sample_id)-tuple is unique. Not sure if
duplicated gene expression results will matter downstream ...
}

