% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-covariates.R
\name{fetch_sample_covariates.FacileDataSet}
\alias{fetch_sample_covariates.FacileDataSet}
\title{Fetch rows from sample_covariate table for specified samples and covariates}
\usage{
\method{fetch_sample_covariates}{FacileDataSet}(x, samples = NULL,
  covariates = NULL, custom_key = Sys.getenv("USER"))
}
\arguments{
\item{samples}{a samples descriptor \code{tbl_*}}

\item{covariates}{character vector of covariate names}

\item{custom_key}{The key to use to fetch more custom annotations over
the given samples}

\item{db}{a \code{FacileDataSet} connection}
}
\value{
rows from the \code{sample_covariate} table
}
\description{
Fetch rows from sample_covariate table for specified samples and covariates
}
\seealso{
Other API: \code{\link{fetch_assay_data.FacileDataSet}},
  \code{\link{fetch_assay_score.FacileDataSet}},
  \code{\link{fetch_custom_sample_covariates.FacileDataSet}},
  \code{\link{fetch_sample_statistics.FacileDataSet}},
  \code{\link{fetch_samples.FacileDataSet}},
  \code{\link{filter_features.FacileDataSet}},
  \code{\link{filter_samples.FacileDataSet}},
  \code{\link{organism.FacileDataSet}},
  \code{\link{samples.FacileDataSet}}
}
\concept{API}
