% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-covariates.R
\name{fetch_sample_covariates}
\alias{fetch_sample_covariates}
\title{Fetch rows from sample_covariate table for specified samples and covariates}
\usage{
fetch_sample_covariates(x, samples = NULL, covariates = NULL,
  custom_key = Sys.getenv("USER"))
}
\arguments{
\item{samples}{a samples descriptor \code{tbl_*}}

\item{covariates}{character vector of covariate names}

\item{custom_key}{The key to use to fetch more custom annotations over
the given samples}

\item{db}{a \code{FacileDataSet} connection}
}
\value{
rows from the \code{sample_covariate} table
}
\description{
Fetch rows from sample_covariate table for specified samples and covariates
}
\seealso{
Other API: \code{\link{fetch_assay_data}},
  \code{\link{fetch_assay_score}},
  \code{\link{fetch_custom_sample_covariates}},
  \code{\link{fetch_sample_statistics}},
  \code{\link{fetch_samples}},
  \code{\link{filter_features}}
}
\concept{API}
