% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{fetch_sample_statistics.default}
\alias{fetch_sample_statistics.default}
\title{Issue #2}
\usage{
\method{fetch_sample_statistics}{default}(x, samples = NULL, semi = TRUE, assay_name = NULL)
}
\description{
Issue #2
}
