% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-info.R
\name{fetch_sample_statistics}
\alias{fetch_sample_statistics}
\title{Fetch the sample statistics for sets of samples in the warehouse}
\usage{
fetch_sample_statistics(x, samples = NULL, semi = TRUE)
}
\arguments{
\item{x}{A \code{FacileDataSet} object}

\item{samples}{a data.frame or tbl_sqlite that has dataset and sample_id
columns}
}
\value{
a tbl_df or tbl_sqlite result from the sample_stats table
}
\description{
Fetch the sample statistics for sets of samples in the warehouse
}
