% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{fetch_samples.FacileDataSet}
\alias{fetch_samples.FacileDataSet}
\title{Fetches a sample descriptor that matches the filter criterion}
\usage{
\method{fetch_samples}{FacileDataSet}(x, samples = NULL, assay = "rnaseq", ...)
}
\arguments{
\item{x}{A \code{FacileDataRepository}}

\item{...}{the NSE boolean filter criteria}
}
\value{
a facile sample descriptor
}
\description{
Use \code{...} as if this is a dplyr::filter call, and our
sample_covariate_tbl was "wide".
}
\details{
This is experimental, so each "term" in the filter criteria should be
just one boolean operation. Multiple terms passed into \code{...} will be
"AND"ed together.
}
\seealso{
Other API: 
\code{\link{fetch_assay_score.FacileDataSet}()},
\code{\link{fetch_custom_sample_covariates.FacileDataSet}()},
\code{\link{fetch_sample_covariates}()},
\code{\link{fetch_sample_statistics.FacileDataSet}()},
\code{\link{filter_features.FacileDataSet}()},
\code{\link{filter_samples.FacileDataSet}()},
\code{\link{organism.FacileDataSet}()},
\code{\link{samples.FacileDataSet}()}
}
\concept{API}
