% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-covariates.R
\name{fetch_samples}
\alias{fetch_samples}
\title{Fetches a sample descriptor that matches filter criterion over covariates.}
\usage{
fetch_samples(db, ...)
}
\arguments{
\item{db}{A \code{FacileDb}}

\item{...}{filter clause to apply to \code{sample_covariate_tbl(db)}}
}
\value{
a \code{tbl} sample-descriptor with a \code{FacileDb} connection
}
\description{
Fetches a sample descriptor that matches filter criterion over covariates.
}
\examples{
fetch_samples(db, indication \%in\% c('BRCA', 'COAD'))
}

