% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay-support.R
\name{filter_by_assay_support}
\alias{filter_by_assay_support}
\title{Removes samples without specific assay support}
\usage{
filter_by_assay_support(x, assay_name, ...)
}
\arguments{
\item{x}{a facile_frame}

\item{assay_name}{the name of the assay}
}
\description{
Samples that do not have data from \code{assay_name} are dropped, and can
retrieved by \code{samples(out, dropped = TRUE)}
}
