% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSE-filter-features.R
\name{filter_features.FacileDataSet}
\alias{filter_features.FacileDataSet}
\title{Filter against the sample_covariate_tbl as if it were wide.}
\usage{
\method{filter_features}{FacileDataSet}(x, ...)
}
\arguments{
\item{x}{A \code{FacileDataSet}}

\item{...}{NSE claused to use in \code{\link[dplyr]{filter}} expressions}
}
\description{
This feature is only really meant to be used
interactively, and with extreme caution ... programatically specifying
column names in feature table, for instance, does not work right now.
}
\details{
TODO: Use tidyeval
}
\seealso{
Other API: 
\code{\link{fetch_assay_score.FacileDataSet}()},
\code{\link{fetch_custom_sample_covariates.FacileDataSet}()},
\code{\link{fetch_sample_covariates}()},
\code{\link{fetch_sample_statistics.FacileDataSet}()},
\code{\link{fetch_samples.FacileDataSet}()},
\code{\link{filter_samples.FacileDataSet}()},
\code{\link{organism.FacileDataSet}()},
\code{\link{samples.FacileDataSet}()}
}
\concept{API}
