% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{filter_samples}
\alias{filter_samples}
\title{Filter x down to specific samples}
\usage{
filter_samples(x, samples = NULL)
}
\arguments{
\item{x}{something like a \code{tbl_sqlite} object}

\item{samples}{a sample descriptor}
}
\value{
filtered version of \code{x} that only has the desired samples
}
\description{
Filter x down to specific samples
}

