% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{flog_level}
\alias{flog_level}
\title{Retrieves the currently set logging level}
\usage{
flog_level(namespace = NULL)
}
\arguments{
\item{namespace}{Package (or whoever) can provide a value here to set the
level they want to listen to. If this is `NULL` (default), the top level
`facile.log.level` value will be used.}
}
\value{
the logging level, as an integer (from `FacileData:::.flog_levels`)
}
\description{
Retrieves the currently set logging level
}
