% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-and-table-functions.R
\name{gene_info_tbl.FacileDataSet}
\alias{gene_info_tbl.FacileDataSet}
\title{Mimics the old \code{gene_info} table.}
\usage{
\method{gene_info_tbl}{FacileDataSet}(x)
}
\description{
Mimics the old \code{gene_info} table.
}
