% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FacileDataSet.R
\name{gene_info_tbl}
\alias{gene_info_tbl}
\title{Mimics the old `gene_info` table.}
\usage{
gene_info_tbl(x)
}
\description{
This function needs to be removed and the code that relies on gene_info_tbl
be update.
}
