% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{fdb}
\alias{fdb}
\alias{fdb<-}
\alias{fdb<-.data.frame}
\alias{fdb<-.tbl}
\alias{set_fdb}
\title{Get/set db}
\usage{
fdb(x)

fdb(x) <- value

\method{fdb}{tbl}(x) <- value

\method{fdb}{data.frame}(x) <- value

set_fdb(x, value)
}
\arguments{
\item{x}{the object}

\item{db}{The \code{FacileDb} object}
}
\description{
Get/set db
}

