% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse-facile-functions.R
\name{group_map.}
\alias{group_map.}
\title{A version of group_map that provides a named list in return}
\usage{
group_map.(.data, .f, ..., .keep = TRUE)
}
\arguments{
\item{.data}{a grouped tibble}

\item{.f}{a function of formula to apply to each group.
If a function, it is used as is. It should have at least 2 formal arguments.
If a formula, e.g. ~ head(.x), it is converted to a function.
In the formula, you can use:
*  \code{.} or \code{.x} to refer to the subset of rows of \code{.tbl} for the given
group
* \code{.y} to refer to the key, a one row tibble with one column per grouping
variable that identifies the group}
}
\value{
a list of elemnts returned by \code{.f} over the grouped elements in
\code{.data}
}
\description{
dplyr::group_map returns an unnamed list, but a named list is often handy
https://github.com/tidyverse/dplyr/issues/4223#issuecomment-469269857
}
\details{
We have also changed the default value for \code{.keep} to be \code{TRUE}, since
I find that almost always use it this way anway.
}
\examples{
# no names
iris |> group_by(Species) |> group_map(~ nrow(.x))

# with names
iris |> group_by(Species) |> group_map.(~ nrow(.x))
}
