% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse-facile-functions.R
\name{group_split.}
\alias{group_split.}
\title{Split a grouped tbl, but return a named list.}
\usage{
group_split.(.tbl, ..., .keep = TRUE)
}
\arguments{
\item{.tbl}{a tibble to split}

\item{...}{the \code{...} in \code{\link[dplyr:group_by]{dplyr::group_by()}}}
}
\value{
a list of the splitted tbl, with named elements
}
\description{
\code{\link[dplyr:group_split]{dplyr::group_split()}} splits a grouped tbl, but the splits are unnamed.
This spanks the names on there.
https://github.com/tidyverse/dplyr/issues/4223#issuecomment-469269857
}
\examples{
# nonames
isplit <- dplyr::group_split(iris, Species)
# names
dsplit <- group_split.(iris, Species)
gsplit <- iris |> group_by(Species) |> group_split.()
}
