% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{has_feature_type}
\alias{has_feature_type}
\title{Test if a given feature type is stored in a FacileDataSet}
\usage{
has_feature_type(x, feature_type)
}
\arguments{
\item{x}{A \code{FacileDataSet}}

\item{feature_type}{a character vector of potential feature types}
}
\value{
logical vector indicating whether or not a given \code{feature_type}
  is stored in \code{x}
}
\description{
Test if a given feature type is stored in a FacileDataSet
}
