% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{hdf5_sample_indices}
\alias{hdf5_sample_indices}
\alias{hdf5_gene_indices}
\title{Utility functions to get row and column indices of rnaseq hdf5 files.}
\usage{
hdf5_sample_indices(x, assay_name = "rnaseq", samples = NULL)

hdf5_gene_indices(x, assay_name = "rnaseq", feature_ids = NULL)
}
\description{
Utility functions to get row and column indices of rnaseq hdf5 files.
}
