% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-types.R
\name{infer_feature_type}
\alias{infer_feature_type}
\title{Guesses the type of feature identifiers from a character vector.}
\usage{
infer_feature_type(x, with_organism = FALSE, ...)
}
\arguments{
\item{x}{a character vector of ids}
}
\value{
data.frame with \code{id} (\code{x}) and \code{id_type}. If \code{with_organism = TRUE},
a third \code{organism} column is added with a guess for the organism.
}
\description{
We rely on meta-information about our data types than "usual", and its useful
to know what types of identifiers we are using for different assay. This
function tries to guess whether an identifier is an ensembl gene identifier,
entrez id, etc.
}
\details{
A two-column data.frame is returned for id_type and organism. Organism
is "unknown" for identifiers where there this can't be inferred (like Refseq).

If an identifier matches more than one id_type, the id_type is set to
\code{"ambiguous"}. If the identifier doesn't match any guesses, then \code{"unknown"}.
}
\examples{
fids <- c("NC_000023", "ENSG00000101811", "ENSMUSG00000030088.2", "85007")
infer_feature_type(fids)
}
