% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construction.R
\name{initializeFacileDataSet}
\alias{initializeFacileDataSet}
\title{Create an empty FacileDataSet}
\usage{
initializeFacileDataSet(path, covariate_definition, page_size = 2^12,
  cache_size = 2e+05)
}
\description{
Create an empty FacileDataSet
}
