% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-db.R
\name{initializeWithExpressionData}
\alias{initializeWithExpressionData}
\title{Creates the expression and sample_stats table, initialize with all data.}
\usage{
initializeWithExpressionData(db, datasets)
}
\arguments{
\item{db}{the db connection}

\item{datasets}{a list of expression sets to load up databases with}
}
\description{
Creates the expression and sample_stats table, initialize with all data.
}
