% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{is.categorical}
\alias{is.categorical}
\title{Tests if a vector is "categorical"}
\usage{
is.categorical(x)
}
\arguments{
\item{x}{a vector to test}
}
\value{
\code{TRUE} if \code{x} is categorical
}
\description{
Is this thing a character or factor?
}
