% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FacileDataSet.R
\name{is.FacileDataSet}
\alias{is.FacileDataSet}
\title{Class and validity checker for FacileDataSet}
\usage{
is.FacileDataSet(x)
}
\arguments{
\item{x}{object to test}
}
\value{
\code{TRUE}/\code{FALSE} indicating that \code{x} nominally "looks like" a
\code{FacileDataSet}
}
\description{
Class and validity checker for FacileDataSet
}
\seealso{
Other FacileDataSet: \code{\link{FacileDataSet}},
  \code{\link{dbfn}}, \code{\link{hdf5fn}},
  \code{\link{meta_file}}
}
\concept{FacileDataSet}
