% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{is.integerish}
\alias{is.integerish}
\title{Tests if a vector is "integerish"}
\usage{
is.integerish(x)
}
\arguments{
\item{x}{a vector to test}
}
\value{
\code{TRUE} if \code{x} is integerish.
}
\description{
Is this thing integer like? As in, is it an integer, or a numeric that has
no decimals?
}
