% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{join_samples}
\alias{join_samples}
\title{Filters the samples down in a dataset to ones specified}
\usage{
join_samples(x, samples = NULL, semi = FALSE,
  distinct.samples = FALSE)
}
\arguments{
\item{x}{likely a \code{tbl_sqlite} object, but a \code{tbl_df}-like
object should work as well.}

\item{samples}{a sample descriptor \code{tbl_df}-like object (likely a
\code{tbl_sqlite} object) that has \code{"dataset"} and \code{"samle_id"}
columns.}

\item{semi}{if \code{TRUE}, appropximates a semi-join on the \code{samples},
otherwise does an inner_join between \code{x} and \code{samples}
(default \code{FALSE}).}
}
\value{
joined result between \code{x} and \code{samples}
}
\description{
Tables like \code{expression} and \code{sample_covariate} house different
datapoints per sample, and we often want to only retreive data points over
a subset of samples.
}
