% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mcast}
\alias{mcast}
\title{Like dcast but the first row becomes rownames of a matrix object}
\usage{
mcast(data, formula, fun.aggregate = NULL, ...)
}
\description{
Like dcast but the first row becomes rownames of a matrix object
}

