% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FacileDataSet.R
\name{meta_file}
\alias{meta_file}
\alias{meta_info}
\alias{organism}
\alias{default_assay}
\alias{dataset_definitions}
\title{Path to the meta information YAML file}
\usage{
meta_file(x)

meta_info(x, fn = meta_file(x))

organism(x)

default_assay(x)

dataset_definitions(x, as.list = TRUE)
}
\arguments{
\item{x}{A \code{FacileDataSet}}

\item{fn}{The path to the \code{meta.yaml} file.}

\item{as.list}{boolean, if \code{FALSE} (default) returns a list, otherwise
sumarraizes results into a tibble.}
}
\value{
The \code{meta.yaml} file parsed into a list-of-lists representation
}
\description{
Path to the meta information YAML file

Get meta information for dataset

Retrieves URL and description of datasets in a FacileDataSet
}
\seealso{
Other FacileDataSet: \code{\link{FacileDataSet}},
  \code{\link{dbfn}}, \code{\link{hdf5fn}},
  \code{\link{is.FacileDataSet}}
}
\concept{FacileDataSet}
