% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metayaml.R
\name{metayaml_skeleton}
\alias{metayaml_skeleton}
\title{Creates a skelton meta.yaml file from a list of assay containers.}
\usage{
metayaml_skeleton(x, name = "UnspecifiedFacileDataSet",
  organism = "unspecified", default_assay = "unspecified",
  covariate_def = list(), ...)
}
\arguments{
\item{x}{list of assay containers}
}
\description{
Creates a skelton meta.yaml file from a list of assay containers.
}
