% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{nameit}
\alias{nameit}
\title{Ensures that a vector has names for all elements if it has names for any}
\usage{
nameit(x, ...)
}
\arguments{
\item{x}{an object with names}
}
\value{
\code{x} with all elements either being uniquely named, or NULL
}
\description{
If the vector is not named, it remains that way
}
