% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FacileDataSet.R
\name{organism.FacileDataSet}
\alias{organism.FacileDataSet}
\title{Retrieves the organism the data is defined over}
\usage{
\method{organism}{FacileDataSet}(x)
}
\value{
`"Homo sapiens`", `"Mus musculus"`, etc.
}
\description{
A FacileDataStore is only expected to hold data for one organism.
}
\seealso{
Other API: \code{\link{fetch_assay_data.FacileDataSet}},
  \code{\link{fetch_assay_score.FacileDataSet}},
  \code{\link{fetch_custom_sample_covariates.FacileDataSet}},
  \code{\link{fetch_sample_covariates}},
  \code{\link{fetch_sample_statistics.FacileDataSet}},
  \code{\link{fetch_samples.FacileDataSet}},
  \code{\link{filter_features.FacileDataSet}},
  \code{\link{filter_samples.FacileDataSet}},
  \code{\link{samples.FacileDataSet}}
}
\concept{API}
