% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FacileDataSet.R
\name{organism}
\alias{organism}
\title{Retrieves the organism the data is defined over}
\usage{
organism(x)
}
\value{
\code{"Homo sapiens}", \code{"Mus musculus"}, etc.
}
\description{
A FacileDataStore is only expected to hold data for one organism.
}
\seealso{
Other API: \code{\link{facet_frame}},
  \code{\link{fetch_assay_data}},
  \code{\link{fetch_assay_score}},
  \code{\link{fetch_custom_sample_covariates}},
  \code{\link{fetch_sample_covariates}},
  \code{\link{fetch_sample_statistics}},
  \code{\link{fetch_samples}},
  \code{\link{filter_features}},
  \code{\link{filter_samples}}, \code{\link{samples}}
}
\concept{API}
