% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{parse_sample_criterion}
\alias{parse_sample_criterion}
\title{Creates a filter expression to select samples based on value of a covariate}
\usage{
parse_sample_criterion(variable, value)
}
\arguments{
\item{variable}{the name of the variable to look for in the sample_covariate
\code{variable} column}

\item{value}{\code{character} vector of values for the \code{variable} that
you want your samples to have.}
}
\value{
a
}
\description{
This leverages dplyr's standard (vs non-standard) evaluation mojo. There is
likely a cleaner way to do this, but to be honest I still find the
\code{\link[lazyeval]{interp}} stuff rather confusing
}
\seealso{
\href{https://cran.r-project.org/web/packages/dplyr/vignettes/nse.html}{dplyr non-standard evaluation}
}
