% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.FacileDataSet.R
\name{pdata_metadata}
\alias{pdata_metadata}
\title{Bioc-container specific pData extraction functions}
\usage{
pdata_metadata(x, ...)
}
\arguments{
\item{x}{SummarizedExperiment, ExpressionSet or DGEList}

\item{...}{additional args, ignored for now}
}
\description{
Get metadata on columns of sample info data.frame (label, etc.) for
inclusion in metadata YAML.
not for export
}
