% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.FacileDataSet.R
\name{pdata_metadata.SummarizedExperiment}
\alias{pdata_metadata.SummarizedExperiment}
\title{SummarizedExperiment method}
\usage{
\method{pdata_metadata}{SummarizedExperiment}(x, ...)
}
\description{
SummarizedExperiment method
}
