% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-and-table-functions.R
\name{primary_key}
\alias{primary_key}
\title{Query a table to identify its primary key(s)}
\usage{
primary_key(x, table_name)
}
\arguments{
\item{x}{a \code{FacileDataSet} or \code{SQLiteConnection}}

\item{table_name}{the name of the table to query}
}
\value{
a character vector of primary keys
}
\description{
Query a table to identify its primary key(s)
}
