% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FacileDataSet.R
\name{print.FacileDataSet}
\alias{print.FacileDataSet}
\title{Print a FacileDataSet}
\usage{
\method{print}{FacileDataSet}(x, ...)
}
\arguments{
\item{x}{FacileDataSet}

\item{...}{additional args (ignored)}
}
\description{
Print a FacileDataSet
}
