% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{tidy}
\alias{arrange}
\alias{collect}
\alias{distinct}
\alias{group_by}
\alias{mutate}
\alias{rename}
\alias{select}
\alias{transmute}
\alias{ungroup}
\alias{left_join}
\alias{inner_join}
\alias{semi_join}
\alias{anti_join}
\alias{nest_join}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{broom}{\code{\link[broom:reexports]{tidy}}}

  \item{dplyr}{\code{\link[dplyr:filter-joins]{anti_join}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr:compute]{collect}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr:mutate-joins]{inner_join}}, \code{\link[dplyr:mutate-joins]{left_join}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{nest_join}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{select}}, \code{\link[dplyr:filter-joins]{semi_join}}, \code{\link[dplyr]{transmute}}, \code{\link[dplyr:group_by]{ungroup}}}
}}

