% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression-sqlite.R
\name{rpkmdb.tbl_sqlite}
\alias{rpkmdb.tbl_df}
\alias{rpkmdb.tbl_sqlite}
\title{Cacluate RPKM from a FacileDataSet}
\usage{
\method{rpkmdb}{tbl_sqlite}(x, gene.length = NULL, lib.size = NULL,
  log = FALSE, prior.count = 5, .fds = fds(x), ...)

\method{rpkmdb}{tbl_sqlite}(x, gene.length = NULL, lib.size = NULL,
  log = FALSE, prior.count = 5, .fds = fds(x), ...)
}
\arguments{
\item{x}{an expression-like facile result}

\item{lib.size}{ignored for now, this is fetched from the
\code{FacileDataSet}}

\item{log}{log the result?}

\item{prior.count}{prior.count to add to observed counts}

\item{.fds}{A \code{FacileDataSet} object}
}
\value{
a modified expression-like result with a \code{rpkm} column.
}
\description{
Cacluate RPKM from a FacileDataSet
}

