% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R, R/sample-covariates.R
\name{fetch_sample_covariates}
\alias{fetch_sample_covariates}
\alias{with_sample_covariates}
\alias{fetch_sample_covariates.FacileDataSet}
\alias{fetch_sample_covariates.facile_frame}
\title{Appends covariate columns to a query result}
\usage{
fetch_sample_covariates(
  x,
  covariates = NULL,
  samples = NULL,
  custom_key = Sys.getenv("USER"),
  with_source = FALSE,
  ...
)

with_sample_covariates(
  x,
  covariates = NULL,
  na.rm = FALSE,
  custom_key = Sys.getenv("USER"),
  .fds = NULL,
  ...
)

\method{fetch_sample_covariates}{FacileDataSet}(
  x,
  covariates = NULL,
  samples = NULL,
  custom_key = Sys.getenv("USER"),
  with_source = FALSE,
  ...
)

\method{fetch_sample_covariates}{facile_frame}(
  x,
  covariates = NULL,
  samples = NULL,
  custom_key = Sys.getenv("USER"),
  with_source = FALSE,
  ...
)
}
\arguments{
\item{x}{a \code{FacileDataSet} connection}

\item{covariates}{character vector of covariate names}

\item{samples}{a samples descriptor \code{tbl_*}}

\item{custom_key}{The key to use to fetch more custom annotations over
the given samples}

\item{na.rm}{if \code{TRUE}, filters outgoing result such that only rows
with nonNA values for the \code{covariates} specified here will be
returned. Default: \code{FALSE}. Note that this will not check columns
not specified in \code{covariates} for NA-ness.}

\item{.fds}{A \code{FacileDataSet} object}
}
\value{
The facile \code{x} object, annotated with the specified covariates.

rows from the \code{sample_covariate} table
}
\description{
Note that this function will force the collection of \code{x}
}
\examples{
# Retrieves a long EAV table of covariates
long_covs <- an_fds() |> 
  fetch_sample_covariates(c("cell_abbrev", "sex", "condition"))
lc2 <- some_samples() |> 
  fetch_sample_covariates(c("cell_abbrev", "sex", "condition"))
wide_covs <- some_samples() |> 
  with_sample_covariates(c("cell_abbrev", "sex", "condition"))
}
\seealso{
Other FacileInterface: 
\code{\link{facet_frame.FacileDataSet}()},
\code{\link{fetch_assay_score}()},
\code{\link{fetch_sample_statistics}()},
\code{\link{samples}()}

Other API: 
\code{\link{fetch_assay_score.FacileDataSet}()},
\code{\link{fetch_custom_sample_covariates.FacileDataSet}()},
\code{\link{fetch_sample_statistics.FacileDataSet}()},
\code{\link{fetch_samples.FacileDataSet}()},
\code{\link{filter_features.FacileDataSet}()},
\code{\link{filter_samples.FacileDataSet}()},
\code{\link{organism.FacileDataSet}()},
\code{\link{samples.FacileDataSet}()}

Other FacileInterface: 
\code{\link{facet_frame.FacileDataSet}()},
\code{\link{fetch_assay_score}()},
\code{\link{fetch_sample_statistics}()},
\code{\link{samples}()}

Other API: 
\code{\link{fetch_assay_score.FacileDataSet}()},
\code{\link{fetch_custom_sample_covariates.FacileDataSet}()},
\code{\link{fetch_sample_statistics.FacileDataSet}()},
\code{\link{fetch_samples.FacileDataSet}()},
\code{\link{filter_features.FacileDataSet}()},
\code{\link{filter_samples.FacileDataSet}()},
\code{\link{organism.FacileDataSet}()},
\code{\link{samples.FacileDataSet}()}

Other API: 
\code{\link{fetch_assay_score.FacileDataSet}()},
\code{\link{fetch_custom_sample_covariates.FacileDataSet}()},
\code{\link{fetch_sample_statistics.FacileDataSet}()},
\code{\link{fetch_samples.FacileDataSet}()},
\code{\link{filter_features.FacileDataSet}()},
\code{\link{filter_samples.FacileDataSet}()},
\code{\link{organism.FacileDataSet}()},
\code{\link{samples.FacileDataSet}()}
}
\concept{API}
\concept{FacileInterface}
