% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facile_interface_generics.R
\name{sample_covariates}
\alias{sample_covariates}
\title{Fetches summary information about sample covariates}
\usage{
sample_covariates(x, samples = NULL, ...)
}
\description{
Returns a table with variable,class,nsamples columns
}
\seealso{
Other FacileInterface: \code{\link{covariate_summary.FacileDataSet}},
  \code{\link{facet_frame.FacileDataSet}},
  \code{\link{organism}}
}
\concept{FacileInterface}
