% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FacileDataSet.R
\name{sample_stats_tbl}
\alias{sample_stats_tbl}
\title{Mimics old sample_stats table}
\usage{
sample_stats_tbl(x)
}
\description{
This function needs to be removed and the code that relies on
sample_stats_tbl be updated.
}
