% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FacileDataSet.R
\name{samples}
\alias{samples}
\title{Retrieves the sample identifiers for all samples in a FacileDataSet.}
\usage{
samples(x)
}
\arguments{
\item{x}{a `FacileDataSet`}
}
\value{
tibble of sample attributes
}
\description{
Sample identifiers are provided as `dataset,sample_id tuples`.
}
\seealso{
Other API: \code{\link{facet_frame}},
  \code{\link{fetch_assay_data}},
  \code{\link{fetch_assay_score}},
  \code{\link{fetch_custom_sample_covariates}},
  \code{\link{fetch_sample_covariates}},
  \code{\link{fetch_sample_statistics}},
  \code{\link{fetch_samples}},
  \code{\link{filter_features}},
  \code{\link{filter_samples}}, \code{\link{organism}}
}
\concept{API}
