% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{some_celltypes}
\alias{some_celltypes}
\title{A set of cell types that exhibit strong expression variability.}
\usage{
some_celltypes()
}
\description{
The \code{\link[=testing_features]{testing_features()}} show cell type expression patterns across these
celltypes
}
