% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{some_features}
\alias{some_features}
\title{A set of features that show variable expression across some_celltypes()}
\usage{
some_features()
}
\description{
These features had strong ANOVA hits across \code{\link[=some_celltypes]{some_celltypes()}}
}
