% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{some_samples}
\alias{some_samples}
\title{Retrieve a set of samples for testing}
\usage{
some_samples(sparse = FALSE)
}
\arguments{
\item{sparse}{If \code{FALSE} (default), a set of samples are returned that have
assay data from both assays in the dataset, otherwise some samples will
be missing scRNAseq data.}
}
\description{
Retrieve a set of samples for testing
}
