% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay-data.R
\name{spread_assay_data}
\alias{spread_assay_data}
\title{Takes a result from fetch_expression and spreads out genes acorss columns}
\usage{
spread_assay_data(x, assay_name, key = c("symbol", "feature_id"),
  .fds = fds(x))
}
\arguments{
\item{x}{facile expression result from \code{fetch_expression}}

\item{key}{the column from the long-form \code{fetch_expression} table
to put in the columns of the outgoing data.frame that the values are
"spread into"}

\item{.fds}{the \code{FacileDataSet}}

\item{value}{the value column to spread into the \code{key} columns}
}
\value{
a more stout \code{x} with the expression values spread across
  columns.
}
\description{
This is a convenience function, and will try to guess what you mean if you
don't explicitly specify which columns to spread and what to call them.
With that mind set, if we find a cpm and symbol column, we will use them
because those are the thing you will likely want to use for exploratory
data analysis if they're in the incoming dataset. If those columns aren't
found, then we'll pick the feature_id and count column.
}
