% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-covariates.R
\name{spread_covariates}
\alias{spread_covariates}
\title{Spreads the covariates returned from database into wide data.frame}
\usage{
spread_covariates(x, cov.def = NULL)
}
\arguments{
\item{x}{output from \code{fetch_sample_covariates}}
}
\value{
a wide \code{tbl_df}-like object
}
\description{
Spreads the covariates returned from database into wide data.frame
}

