% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{spread_expression}
\alias{spread_expression}
\title{Takes a result from fetch_expressoi and spreads out genes acorss columns}
\usage{
spread_expression(x, key = c("symbol", "feature_id"), value = c("cpm",
  "count"), .fds = fds(x))
}
\arguments{
\item{x}{facile expression result from \code{fetch_expression}}
}
\value{
a more stout \code{x} with the expression values spread across
  columns.
}
\description{
This is a convenience function, and will try to guess what you mean if you
don't explicitly specify which columns to spread and what to call them.
With that mind set, if we find a cpm and symbol column, we will use them
because those are the thing you will likely want to use for exploratory
data analysis if they're in the incoming dataset. If those columns aren't
found, then we'll pick the feature_id and count column.
}

