% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{sqlFromFile}
\alias{sqlFromFile}
\title{Utility function to send more than one sql command to the database}
\usage{
sqlFromFile(file)
}
\arguments{
\item{file}{single character, name of file with SQL statements}
}
\description{
Copied from http://stackoverflow.com/questions/18914283
}
