% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{subtype_map}
\alias{subtype_map}
\title{Creates tbl that associates a sample with all of its indication/subtypes}
\usage{
subtype_map(db)
}
\arguments{
\item{db}{a \code{FacileDb} objectg}
}
\value{
a \code{tbl} with indication and subtype information for all samples
  in the database
}
\description{
Creates tbl that associates a sample with all of its indication/subtypes
}

