% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-covariates.R
\name{summary.eav_covariates}
\alias{summary.eav_covariates}
\title{Provides a summary table of sample covariates.}
\usage{
\method{summary}{eav_covariates}(object, expanded = FALSE, droplevels = TRUE, ...)
}
\arguments{
\item{object}{A sample covariate table, the likes returned from
\code{\link[=fetch_sample_covariates]{fetch_sample_covariates()}}.}

\item{expanded}{includes details (rows) for each covariate per level
(or quantile), depending on the covariates \code{"class"} attribute.}
}
\value{
a tibble of summary sample covariate information with the following
columns:
\itemize{
\item \code{variable}: name of the variable
\item \code{class}: class of variable (real, categorical)
\item \code{nsamples}: the number of samples that have this variable defined
\item \code{level}: the level (or quantile) of the covariate
(included only when \code{expanded == TRUE})
\item \code{ninlevel}: the number of samples with this covariate value
(included only when \code{expanded == TRUE})
}
}
\description{
Sumamrizes a set of sample covariates (returned from
\code{\link[=fetch_sample_covariates]{fetch_sample_covariates()}} at different granulaities.
}
\examples{
fds <- exampleFacileDataSet()
covs <- fetch_sample_covariates(fds)
smry <- summary(covs)
details <- summary(covs, expanded = TRUE)
catdeetz <- covs |>
  filter(class == "categorical") |>
  summary(expanded = TRUE)
}
