% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{TestDb}
\alias{TestDb}
\title{Creates a FacileDb connection to a test database}
\usage{
TestDb(db.type = c("monetdblite", "sqlite"),
  datapath = getOption("ftest.datapath", NULL),
  db.fn = getOption("ftest.dbpath", NULL),
  covdef.fn = getOption("ftest.covdef"))
}
\arguments{
\item{db.fn}{The path on the filesystem to the facile database}

\item{covdef}{The path on the filesystem to the yaml file that provides
the covariate definitions}
}
\description{
TODO: This function currently uses the Atezo database as a test. We need to
      create a REAL testing database.
}

