% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{testFacileDataSet}
\alias{testFacileDataSet}
\title{Retrieves a multi-modal FacileDataSet based on KPMP data}
\usage{
testFacileDataSet()
}
\value{
a FacileDataSet
}
\description{
Retrieves a multi-modal FacileDataSet based on KPMP data
}
