% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-db.R
\name{updateSampleCovariates}
\alias{updateSampleCovariates}
\title{Update or add new sample covariate annotation}
\usage{
updateSampleCovariates(db, covariates)
}
\arguments{
\item{db}{a \code{FacileDb} object}

\item{covariates}{the covariate table to insert/replace}
}
\description{
Update or add new sample covariate annotation
}

