% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity-attribute-value.R
\name{validate_covariate_def_list}
\alias{validate_covariate_def_list}
\title{Validates that a covariate defintion list reasonably describes a data.frame}
\usage{
validate_covariate_def_list(x, pdata)
}
\arguments{
\item{x}{a covariate definition list-of-lists}

\item{pdata}{a \code{data.frame}}
}
\description{
Validates that a covariate defintion list reasonably describes a data.frame
}
