% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay-data.R
\name{with_assay_data.facile_frame}
\alias{with_assay_data.facile_frame}
\title{Append expression values to sample-descriptor}
\usage{
\method{with_assay_data}{facile_frame}(x, features, assay_name = NULL,
  normalized = TRUE, aggregate.by = NULL, spread = TRUE,
  with_assay_name = FALSE, ..., verbose = FALSE, .fds = fds(x))
}
\arguments{
\item{.fds}{A \code{FacileDataSet} object}

\item{samples}{a samples descriptor}

\item{feature_ids}{character vector of feature_ids}

\item{with_symbols}{Do you want gene symbols returned, too?}
}
\value{
a tbl-like result
}
\description{
Since this is called in a "convenience" sort of way, often in a pipe-chain
\code{normalize} defaults to \code{TRUE}
}
