% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay-data.R
\name{with_assay_data}
\alias{with_assay_data}
\title{Append expression values to sample-descriptor}
\usage{
with_assay_data(samples, features, assay_name = NULL, normalized = FALSE,
  aggregate.by = NULL, spread = NULL, ..., .fds = fds(samples))
}
\arguments{
\item{.fds}{A \code{FacileDataSet} object}

\item{x}{a samples descriptor}

\item{feature_ids}{character vector of feature_ids}

\item{with_symbols}{Do you want gene symbols returned, too?}
}
\value{
a tbl-like result
}
\description{
Append expression values to sample-descriptor
}
