% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression-sqlite.R
\name{with_expression.db}
\alias{with_expression.db}
\title{Append expression values to sample-descriptor}
\usage{
with_expression.db(samples, feature_ids, with_symbols = TRUE,
  .fds = fds(samples))
}
\arguments{
\item{feature_ids}{character vector of feature_ids}

\item{with_symbols}{Do you want gene symbols returned, too?}

\item{.fds}{A \code{FacileDataSet} object}

\item{x}{a samples descriptor}
}
\value{
a tbl-like result
}
\description{
Append expression values to sample-descriptor
}
