% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{with_expression}
\alias{with_expression}
\title{Append expression values to sample-descriptor}
\usage{
with_expression(samples, feature_ids, db = fdb(samples))
}
\arguments{
\item{feature_ids}{character vector of feature_ids}

\item{x}{a samples descriptor}

\item{a}{\code{FacileDb} object}
}
\value{
a tbl-like result
}
\description{
Append expression values to sample-descriptor
}

