% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facile_interface_generics.R
\name{with_feature_info}
\alias{with_feature_info}
\title{Append feature information columns to (feature-rows)}
\usage{
with_feature_info(x, covariates = NULL, ...)
}
\arguments{
\item{x}{a data.frame feature descriptor columns (feature_id, feature_type)}
}
\value{
`x` fattened with the columns asked for
}
\description{
Append feature information columns to (feature-rows)
}
