% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-covariates.R
\name{with_sample_covariates}
\alias{with_sample_covariates}
\title{Appends covariate columns to a query result}
\usage{
with_sample_covariates(x, covariates = NULL, db = fdb(x),
  cov.def = db[["cov.def"]])
}
\arguments{
\item{x}{a "tbl"-like object that is a sample descriptor}
}
\description{
Note that this function will force the collection of \code{x}
}

