% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construction.R
\name{addFacileAssaySet}
\alias{addFacileAssaySet}
\title{Adds a complete set of assay data for all samples across datasets in the
FacileDataSet}
\usage{
addFacileAssaySet(x, datasets, facile_assay_name,
  facile_assay_type = .assay.types, facile_feature_type = .feature.types,
  facile_assay_description = NULL, facile_feature_info,
  sorage_mode = .storage.modes, chunk_rows = 5000, chunk_cols = "ncol",
  chunk_compression = 4, assay_name = NULL)
}
\arguments{
\item{facile_assay_name}{the name of the assay to store in the FacileDataSet}

\item{facile_assay_type}{string indicating the assay_type}

\item{assay_name}{the name of the assay in the source dataset object}

\item{dat}{list of ExpressionSet, SummarizedExperiment, or DGELists that
have the assay data for the given assay across all of our datasets}
}
\description{
This needs to be busted up into functions Minimally loop over datasets to
addFacileAssay
}
