% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construction.R
\name{append_facile_feature_info}
\alias{append_facile_feature_info}
\title{Appends new features to \code{feature_info} table}
\usage{
append_facile_feature_info(
  x,
  feature_info,
  type = feature_info$feature_type,
  warn_existing = FALSE
)
}
\arguments{
\item{x}{The \code{FacileDataSet}}

\item{feature_info}{a table of new features that provides all columns
in \code{feature_info_tbl(x)}}

\item{type}{A way to override (or set) the \code{feature_type} column of the
\code{feature_info} table}
}
\value{
invisible returns an annotated version of the \code{feature_info}
table with an \verb{$added} column with \code{TRUE/FALSE} values for the
features that were new (and added) to the repository or \code{FALSE} to
indicate that they were already in the database.
}
\description{
This function only adds features (feature_type, feature_id) that are not
in the \code{feature_info} table already
}
