% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construction.R
\name{append_facile_table}
\alias{append_facile_table}
\title{Adds rows to a table in a FacileDataSet}
\usage{
append_facile_table(dat, x, table_name)
}
\arguments{
\item{dat}{the \code{data.frame} of rows to add to the table, which must
have a superset of columns present in the \code{table_name} that is being
appended to}

\item{x}{the \code{FacileDataSet}}

\item{table_name}{the name of the table in \code{x} to add the rows of
\code{dat} to.}
}
\value{
invisibly returns the conformed version of \code{dat}.
}
\description{
This function uses \code{\link{conform_data_frame}} to match up
\code{dat} to \code{table_name}.
}
