% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.BiocAssayContainers.R
\name{as.DGEList}
\alias{as.DGEList}
\alias{as.DGEList.matrix}
\alias{as.DGEList.data.frame}
\alias{as.DGEList.tbl_sql}
\alias{as.ExpressionSet}
\alias{as.ExpressionSet.data.frame}
\alias{as.ExpressionSet.FacileDataSet}
\alias{as.SummarizedExperiment}
\alias{as.SummarizedExperiment.data.frame}
\alias{as.SummarizedExperiment.FacileDataSet}
\title{Converts a facile result into a traditional Bioconductor assay container}
\usage{
as.DGEList(x, ...)

\method{as.DGEList}{matrix}(x, covariates = TRUE, feature_ids = NULL,
  assay_name = default_assay(.fds), .fds = fds(x),
  custom_key = Sys.getenv("USER"), ...)

\method{as.DGEList}{data.frame}(x, covariates = TRUE,
  feature_ids = NULL, assay_name = default_assay(.fds),
  .fds = fds(x), custom_key = Sys.getenv("USER"), ...)

\method{as.DGEList}{tbl_sql}(x, covariates = TRUE, feature_ids = NULL,
  assay_name = default_assay(.fds), .fds = fds(x),
  custom_key = Sys.getenv("USER"), ...)

as.ExpressionSet(x, ...)

\method{as.ExpressionSet}{data.frame}(x, covariates = TRUE,
  feature_ids = NULL, assay_name = default_assay(.fds),
  .fds = fds(x), custom_key = Sys.getenv("USER"), ...)

\method{as.ExpressionSet}{FacileDataSet}(x, covariates = TRUE,
  feature_ids = NULL, assay_name = default_assay(.fds),
  .fds = fds(x), custom_key = Sys.getenv("USER"), ...)

as.SummarizedExperiment(x, ...)

\method{as.SummarizedExperiment}{data.frame}(x, covariates = TRUE,
  feature_ids = NULL, assay_name = default_assay(.fds),
  .fds = fds(x), custom_key = Sys.getenv("USER"), ...)

\method{as.SummarizedExperiment}{FacileDataSet}(x, covariates = TRUE,
  feature_ids = NULL, assay_name = default_assay(.fds),
  .fds = fds(x), custom_key = Sys.getenv("USER"), ...)
}
\arguments{
\item{x}{a facile expression-like result}

\item{covariates}{The covariates the user wants to add to the $samples of
the DGEList. This can take the following forms:
\itemize{
\item \code{TRUE}: All covariates are retrieved from the \code{FacileDataSet}
\item \code{FALSE}: TODO: Better handle FALSE
\item \code{character}: A vector of covariate names to fetch from the
\code{FacileDataSet}
\item \code{data.frame}: A table that looks like a subset of the
\code{sample_covariate}, which will be transformed into the \code{pData}
\item \code{NULL}: do not decorate with \emph{any} covariates.
}}

\item{feature_ids}{the features to get expression for (if not specified
in \code{x} descriptor)}

\item{.fds}{The \code{FacileDataSet} that \code{x} was retrieved from}

\item{custom_key}{the custom key to use to fetch custom annotations from
\code{.fds}}

\item{assay}{the assay matrix to use when populating the default assay
matrix of the bioconductor container, i.e. the \code{$counts} matrix of a
\code{DGEList}, the \code{exprs()} of an \code{ExpressionSet}, etc.}
}
\value{
the appropriate bioconductor assay container, ie. a \link[edgeR:DGEList]{edgeR::DGEList}
for \code{as.DGEList}, an \link[Biobase:ExpressionSet]{Biobase::ExpressionSet} for \code{as.ExpressionSet}, or
a \link[SummarizedExperiment:SummarizedExperiment]{SummarizedExperiment::SummarizedExperiment} for
\code{as.SummarizedExperiment}.

a \code{\link[Biobase]{ExpressionSet}}

a \code{\link[SummarizedExperiment]{SummarizedExperiment}}
}
\description{
Even though the faciledata API provides functionality to access your data
in useful ways, it is still conceivable that you \emph{might} prefer to work
with these data using a more traditional bioconductor container like a
\code{SummarizedExperiment}, \code{DGEList}, \code{ExpressionSet}, etc.
}
