% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.FacileDataSet.R
\name{as.FacileDataSet}
\alias{as.FacileDataSet}
\alias{as.FacileDataSet.ExpressionSet}
\alias{as.FacileDataSet.SummarizedExperiment}
\alias{as.FacileDataSet.DGEList}
\alias{as.FacileDataSet.list}
\title{Converts bioconductor assay containers into a FacileDataSet.}
\usage{
as.FacileDataSet(x, assay_name, assay_type, feature_info, feature_type,
  metayaml = NULL, organism = "unspecified",
  path = tempfile("FacileDataSet-", getwd()), source_assay = NULL, ...)

\method{as.FacileDataSet}{ExpressionSet}(x, assay_name, assay_type,
  feature_info, feature_type, metayaml = NULL, organism = "unspecified",
  path = tempfile("FacileDataSet-", getwd()),
  source_assay = assayDataElementNames(x)[1L], dset = "eset", ...)

\method{as.FacileDataSet}{SummarizedExperiment}(x, assay_name, assay_type,
  feature_info, feature_type, metayaml = NULL, organism = "unspecified",
  path = tempfile("FacileDataSet-", getwd()), source_assay = NULL,
  dset = "sumexp", ...)

\method{as.FacileDataSet}{DGEList}(x, assay_name, assay_type, feature_info,
  feature_type, metayaml = NULL, organism = "unspecified",
  path = tempfile("FacileDataSet-", getwd()), source_assay = NULL,
  dset = "dgelist", ...)

\method{as.FacileDataSet}{list}(x, path, organism, assays = NULL,
  metayaml = NULL, ...)
}
\arguments{
\item{x}{The bioconductor assay container to extract data from}

\item{assay_name}{The name to use when storing the assay matrix from
\code{x} into the faciledataset.}

\item{assay_type}{what type of assay is this? rnaseq, microarry, nanostring,
isoseq (isoform expression), etc.}

\item{feature_info}{a \code{data.frame} that describes the information for the
features (rows) of the assay you are extracting. Currently you had to
hand-craft this. In the future we will provide automated default
fData, rowData, etc. extractors for the source assay cointaner.}

\item{feature_type}{\code{c('entrez', 'ensgid', 'enstid', 'genomic', 'custom')}}

\item{metayaml}{a yaml file (or list of lists) that describes the covariates
in the pData/colData of \code{x}. If not provided, a default one will be
generated}

\item{organism}{This is used to fetch the appropriate genesets when this
dataset is used with the facileexplorer. Currently supported values are:
\code{c("Homo sapiens", "Mus musculus", "unspecified")}.}

\item{path}{the directory to create the faciledataset into. Will create
a default directory in the current working directory if not specified.}

\item{source_assay}{the name of the assay element in \code{x} to extract
for use.}

\item{...}{more args}
}
\value{
a \code{\link[=FacileDataSet]{FacileDataSet()}}
}
\description{
This function assumes you are only extracting one assay from the assay
container and creating a \code{FacileDataSet} from it. This requires that you
specify which assay (if the container has more than one) to extract. The
\code{feature_info} table is populated by the particular \code{fData} (\code{mcols},
\code{y$genes}, etc.) from the assay container. The \code{pData} from the assay
container will be ingestged as well.

\strong{Please read the Details section}  for more a more complete description
of the data that is required to create a complete \code{FacileDataSet}
}
\details{
A \code{FacileDataSet} can be created from a number of different Bioconductor
containers, such as a \link[Biobase:ExpressionSet]{Biobase::ExpressionSet},
\link[SummarizedExperiment:SummarizedExperiment]{SummarizedExperiment::SummarizedExperiment}, or an \link[edgeR:DGEList]{edgeR::DGEList}. To
create a \code{FacileDataSet} that spans multiple Bioc containters, ie. you may
have one ExpressionSet per indication in the TCGA. You can make
\code{FacileDataSet} to encompass the data from \emph{all} of these indications by
providing a \code{list} of \code{ExpressionSet}s. The \code{list} should have its \code{names()}
set to each of the TCGA indications ("BLCA", "BRCA", etc.) the data came
from.
}
