% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facile_frame.R
\name{as_facile_frame}
\alias{as_facile_frame}
\title{Converts a normal tibble/data.frame to a facile_frame}
\usage{
as_facile_frame(x, datastore = fds(x), classes = NULL, ...,
  .valid_sample_check = TRUE)
}
\arguments{
\item{x}{a sample-like descriptor}

\item{datastore}{the FacileDataStore tied to x}

\item{classes}{more classes to append to the outgoing object. The
`"facile_frame"` class entry is always the last one of the bunch.}

\item{...}{dots}

\item{.valid_sample_check}{If `TRUE` (default), will check if `x` is a valid
subset of the FacileDataStore `.fds`. Internal functions may set this to
`TRUE` to avoid the check to (1) save time; and (2) save infinite
recursion in the call to `assert_sample_subset`.}
}
\description{
Converts a normal tibble/data.frame to a facile_frame
}
