% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay-data.R
\name{assay_feature_info}
\alias{assay_feature_info}
\title{Materializes a table with all feature information for a given assay.}
\usage{
assay_feature_info(x, assay_name, feature_ids = NULL)
}
\arguments{
\item{x}{\code{FacileDataSet}}

\item{assay_name}{the name of the assay}

\item{feature_ids}{a character vector of feature_ids}
}
\value{
a \code{tbl_sqlite} result with the feature information for the
features in a specified assay
}
\description{
DEBUG: This logic is unnecessarily complex because I make sure to collect
all tables from the database as opposed to copying external tables in and
doing an inner_join in the database. I'm doing this becuase we are getting
name collections on some of the temporary tables. We get errors like:
Warning: Error in : Table pkdtpohpsu already exists.
}
\details{
This fetches the hdf5_index for the assays as well
}
