% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facile_interface_generics.R
\name{assay_info}
\alias{assay_info}
\title{The Facile API}
\usage{
assay_info(x, assay_name = NULL, ...)
}
\arguments{
\item{x}{A \code{FacileDataStore}}

\item{assay_name}{optional name of the assay to get information for}
}
\value{
a tibble of meta information for the assays stored in \code{x}
}
\description{
This is a stub for a manpage all about the FacileAPI
Fetches assay meta information for the assays stored in a FacileDataStore
}
