% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facile_interface_generics.R
\name{assay_units}
\alias{assay_units}
\title{The Facile API}
\usage{
assay_units(x, assay_name, normalized = FALSE, abbreviate = FALSE, ...)
}
\value{
string
}
\description{
This is a stub for a manpage all about the FacileAPI
Units of measure in an assay
}
