% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{biocbox}
\alias{biocbox}
\title{Materialize a Bioconductor assay container from some facile object.}
\usage{
biocbox(x, ...)
}
\arguments{
\item{x}{A facile object}
}
\description{
Most often, this will be from some facile_frame to create a Bioconductor
assay container object, but this function can be overloaded for other
purposes.
}
\details{
The FacileAnalysis package, for example, uses this function to materialize
bioconductor objects of different flavors from different analysis results,
ie. a DGEList, or perhaps a limma fit object, etc.
}
