% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-covariates.R
\name{cast_covariate}
\alias{cast_covariate}
\title{Casts the character values of the covariates to their defined types}
\usage{
cast_covariate(covariate, values, cov.defs)
}
\arguments{
\item{covariate}{the name of the covariate}

\item{values}{the covariate values (likely as a character)}

\item{cov.defs}{the path to the covariate definition yaml file}
}
\value{
values cast to appropriate type if a valid definition was found for
  \code{covariate}, otherwise values is returned "as is"
}
\description{
Casts the character values of the covariates to their defined types
}

