% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity-attribute-value.R
\name{cast_covariate}
\alias{cast_covariate}
\title{Casts the character EAV values to their R-native defined types.}
\usage{
cast_covariate(covariate, values, cov.def, .fds)
}
\arguments{
\item{covariate}{the name of the covariate}

\item{values}{the covariate values (which is a \code{character}) as it is
pulled from the database.}

\item{cov.def}{the un-yamled covariate definitions, if missing we rely on
pulling this out from the \code{FacileDataSet} object \code{.fds}}

\item{.fds}{If \code{missing(cov.def)}, this is the \code{FacileDataSet} to
get the covariate definitions from.}
}
\value{
values cast to appropriate type if a valid definition was found for
\code{covariate}, otherwise values is returned "as is". Most of the time
this is a single vector, but others it can be a data.frame (for
\code{right_censored} data, for instance)
}
\description{
For most things, a single value will be returned from each cast, but in the
case of "time_to_event" data, the value is expended to a two column
data.frame with a \code{tte_<covariate>} column for time to event, and an
\code{event_<covariate>} column to indicate event (1) or right censored (2).
}
\details{
The mechanics of how values in the \code{sample_covariate} table are converted
into R objects are handled by the information stored in the
\code{FacileDataSets}'s \code{meta.yaml} file.
}
\seealso{
\code{\link[=covariate_meta_info]{covariate_meta_info()}}, \code{\link[=covariae_defnitions]{covariae_defnitions()}}
}
