% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{covariate_definitions}
\alias{covariate_definitions}
\title{Get description of sample metadata columns}
\usage{
covariate_definitions(x, as.list = TRUE, ...)
}
\arguments{
\item{x}{FacileDataTore}

\item{as.list}{single logical, return tibble or list}
}
\value{
meta information about the sample covariates in \code{x}
}
\description{
Descriptions of the sample covariates can be specified in a FacileDataSet's
\code{meta.yaml} file. This function returns those.
}
