% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-db.R
\name{createWarehouse}
\alias{createWarehouse}
\title{Creates the database}
\usage{
createWarehouse(db.path, datasets, sample.meta, pragma.page_size = 2^12,
  db.type = "monetdblite", ...)
}
\description{
Creates the database
}

