% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FacileDataSet.R
\name{dbfn}
\alias{dbfn}
\title{Get location of the FacileDataSet database}
\usage{
dbfn(x, mustWork = TRUE)
}
\arguments{
\item{x}{FacileDataSet}

\item{mustWork}{boolean, if \code{TRUE} (default), throws an error if the sqlite
file does not exist. When \code{FALSE}, this returns the "expected" path to the
sqlite file for \code{x}}
}
\value{
the filepath to the sqlite database
}
\description{
Get location of the FacileDataSet database
}
\seealso{
Other FacileDataSet: 
\code{\link{FacileDataSet}()},
\code{\link{hdf5fn}()},
\code{\link{meta_file}()}
}
\concept{FacileDataSet}
