% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity-attribute-value.R
\name{eav_encode_right_censored}
\alias{eav_encode_right_censored}
\alias{eav_decode_right_censored}
\title{Entity-attribute-value encodings for survival data.}
\usage{
eav_encode_right_censored(time, event, sas.encoding = FALSE, ...)

eav_decode_right_censored(x, attrname = character(), def = list(),
  suffix = attrname, sas.encoding = FALSE, ...)
}
\arguments{
\item{time}{\code{numeric} time to event}

\item{event}{0/1 vector encoded in the "R sense". "1" is an event, "0" is
right censored.}

\item{sas.encoding}{Is the 'event' vector "SAS encoded"? In the SAS world,
1 means censored, and 0 is event. This is \code{FALSE} by default.}

\item{x}{the time to event}

\item{def}{the covariate definition for this variable}

\item{suffix}{adds \code{_<suffix>} to the \code{tte} and \code{event} columns of the
outgoing \code{data.frame}}
}
\value{
returns a numeric vector that combines time-to-event and censoring
info (sign of the value).

two column \code{data.frame} with \code{tte(_SUFFIX)?} and \code{event(_SUFFIX)?}
columns.
}
\description{
Entity-attribute-value encodings for survival data.
}
\details{
Encoding of survival data in R requires two columns, one to store
the time-to-event and another to indicate if there was an "event" at stored
time, or if it was censored. A \code{FacileDataSet} stores these two \code{pData}
columns into one "value" column in its entity-attribute-value
\code{sample_covariate} table.

The \code{encode_right_censored} function takes the time-to-event and censoring
vectors and encodes them into a single signed time-to-event numeric value.
Positive values indicate an event, and negative value are censored.

The \code{decode_right_censored} function re-instantiates the two-column R-native
storage of this data.
}
\seealso{
\code{\link[=eav_metadata_create]{eav_metadata_create()}}
}
