% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity-attribute-value.R
\name{eav_metadata_merge}
\alias{eav_metadata_merge}
\title{Merge inferred and explicit pdata column metadata}
\usage{
eav_metadata_merge(pdat, covariate_def)
}
\arguments{
\item{pdat}{data.frame, from one or more datasets, bind_rows-d together}

\item{covariate_def}{list of additional covariate info, such as 'label'.}
}
\value{
list of column metadata lists
}
\description{
Inferred metadata comes from inspecting the types of each column
of the sample info data frame. Explicit metadata enters
as.FacileDataSet attached to the incoming BioC object and is read
by calling pdata_metadata on that object. The colnames of 'pdat'
must match the names of 'covariate_def'.
}
