% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity-attribute-value.R
\name{eavdef_for_column}
\alias{eavdef_for_column}
\title{Generate entity-attribute-value definition for a column in a data.frame}
\usage{
eavdef_for_column(column, column_name)
}
\arguments{
\item{column}{a vector, e.g. a column out of a pdata}

\item{column_name}{single character, name of the colum}
}
\value{
a generic list-of-list definition column
}
\description{
Creates the minimal list-definition for a single column in a \code{pData}
\code{data.frame}. This function is not exported on purpose. Column descriptions
will be taken from the "label" attribute of data.frames or the "metadata" list
for DataFrames.
}
