% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{exampleFacileDataSet}
\alias{exampleFacileDataSet}
\title{Retrieves an example FacileDataSet}
\usage{
exampleFacileDataSet()
}
\description{
A subset of the TCGA data from the BLCA and COAD indications is provided
as a FacileDataSet.
}
