% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{exampleFacileDataSet}
\alias{exampleFacileDataSet}
\title{Creates a FacileDb connection to a test database}
\usage{
exampleFacileDataSet(db.type = c("sqlite", "monetdblite"))
}
\arguments{
\item{db.fn}{The path on the filesystem to the facile database}

\item{covdef}{The path on the filesystem to the yaml file that provides
the covariate definitions}
}
\description{
TODO: This function currently uses the Atezo database as a test. We need to
      create a REAL testing database.
}
