% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression-sqlite.R, R/expression.R
\name{as.DGEList.db}
\alias{as.DGEList}
\alias{as.DGEList.data.frame}
\alias{as.DGEList.db}
\alias{as.DGEList.matrix}
\alias{as.DGEList.tbl_sqlite}
\alias{as.ExpressionSet}
\alias{as.ExpressionSet.db}
\title{Converts a result from `fetch_expression` into a DGEList}
\usage{
\method{as.DGEList}{db}(x, covariates = NULL, .fds = fds(x), ...)

as.ExpressionSet.db(x, covariates = NULL, exprs = "counts", .fds = fds(x),
  ...)

as.DGEList(x, ...)

\method{as.DGEList}{matrix}(x, covariates = TRUE, feature_ids = NULL,
  .fds = fds(x), custom_key = Sys.getenv("USER"), ...)

\method{as.DGEList}{data.frame}(x, covariates = TRUE, feature_ids = NULL,
  .fds = fds(x), custom_key = Sys.getenv("USER"), ...)

\method{as.DGEList}{tbl_sqlite}(x, covariates = TRUE, feature_ids = NULL,
  .fds = fds(x), custom_key = Sys.getenv("USER"), ...)

as.ExpressionSet(x, covariates = TRUE, feature_ids = NULL,
  exprs = "counts", .fds = fds(x), custom_key = Sys.getenv("USER"), ...)
}
\arguments{
\item{x}{a facile expression-like result}

\item{covariates}{A \code{character} vector specifying the  additional
covariates to append to \code{out$samples}. Must be valid entries in the
\code{sample_covariate::variable} column.}

\item{.fds}{The \code{FacileDataSet} that \code{x} was retrieved from.}

\item{feature_ids}{the features to get expression for (if not specified
in \code{x} descriptor)}

\item{custom_key}{the custom key to use to fetch custom annotations from
\code{.fds}}

\item{x}{a facile expression-like result}

\item{covariates}{A \code{character} vector specifying the  additional
covariates to append to \code{out$samples}. Must be valid entries in the
\code{sample_covariate::variable} column.}

\item{assay}{Which column to put in \code{"exprs"}}

\item{.fds}{The \code{FacileDataSet} that \code{x} was retrieved from.}

\item{x}{a facile expression-like result}

\item{covariates}{The covariates the user wants to add to the $samples of
the DGEList. This can take the following forms:
\enumerate{
  \item{TRUE}{
    All covariates are retrieved from the \code{FacileDataSet}
  }
  \item{character}{
    A vector of covariate names to fetch from the \code{FacileDataSet}
  }
  \item{data.frame}{
    A table that looks like a subset of the sample_covariate table
  }
  \item{NULL}{
    If \code{NULL}, no covariates are retrieved
  }
}}

\item{assay}{the \code{assayDataElement} to use for the expression data
if \code{x} is an \code{ExpressionSet}.}

\item{.fds}{The \code{FacileDataSet} that \code{x} was retrieved from.}
}
\value{
a \code{\link[edgeR]{DGEList}}

a \code{\link[Biobase]{ExpressionSet}}

a \code{\link[edgeR]{DGEList}}
}
\description{
The genes and samples that populate the \code{DGEList} are specified by
\code{x}, and the caller can request addition sample information to be
appended to \code{out$samples} via specification through the
\code{covariates} argument.

Create an ExpressionSet from `fetch_expression`.

The genes and samples that populate the \code{DGEList} are specified by
\code{x}, and the caller can request addition sample information to be
appended to \code{out$samples} via specification through the
\code{covariates} argument.
}

