% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression-sqlite.R, R/expression.R
\name{as.DGEList.db}
\alias{as.DGEList.db}
\alias{as.DGEList.matrix}
\alias{as.ExpressionSet}
\alias{as.ExpressionSet.db}
\title{Converts a result from `fetch_expression` into a DGEList}
\usage{
\method{as.DGEList}{db}(x, covariates = NULL, .fds = fds(x), ...)

as.ExpressionSet.db(x, covariates = NULL, exprs = "counts", .fds = fds(x),
  ...)

\method{as.DGEList}{matrix}(x, covariates = NULL, .fds = fds(x),
  samples = NULL, ...)

as.ExpressionSet(x, covariates = NULL, exprs = "counts", .fds = fds(x),
  ...)
}
\arguments{
\item{x}{a facile expression-like result}

\item{covariates}{A \code{character} vector specifying the  additional
covariates to append to \code{out$samples}. Must be valid entries in the
\code{sample_covariate::variable} column.}

\item{.fds}{The \code{FacileDataSet} that \code{x} was retrieved from.}

\item{x}{a facile expression-like result}

\item{covariates}{A \code{character} vector specifying the  additional
covariates to append to \code{out$samples}. Must be valid entries in the
\code{sample_covariate::variable} column.}

\item{assay}{Which column to put in \code{"exprs"}}

\item{.fds}{The \code{FacileDataSet} that \code{x} was retrieved from.}

\item{x}{a facile expression-like result}

\item{covariates}{A \code{character} vector specifying the  additional
covariates to append to \code{out$samples}. Must be valid entries in the
\code{sample_covariate::variable} column.}

\item{.fds}{The \code{FacileDataSet} that \code{x} was retrieved from.}

\item{x}{a facile expression-like result}

\item{covariates}{A \code{character} vector specifying the  additional
covariates to append to \code{out$samples}. Must be valid entries in the
\code{sample_covariate::variable} column.}

\item{assay}{Which column to put in \code{"exprs"}}

\item{.fds}{The \code{FacileDataSet} that \code{x} was retrieved from.}
}
\value{
a \code{\link[edgeR]{DGEList}}

a \code{\link[Biobase]{ExpressionSet}}

a \code{\link[edgeR]{DGEList}}

a \code{\link[Biobase]{ExpressionSet}}
}
\description{
The genes and samples that populate the \code{DGEList} are specified by
\code{x}, and the caller can request addition sample information to be
appended to \code{out$samples} via specification through the
\code{covariates} argument.

Create an ExpressionSet from `fetch_expression`.

The genes and samples that populate the \code{DGEList} are specified by
\code{x}, and the caller can request addition sample information to be
appended to \code{out$samples} via specification through the
\code{covariates} argument.

Create an ExpressionSet from `fetch_expression`.
}

