% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FacileDataSet.R
\name{facet_frame}
\alias{facet_frame}
\alias{facet_frame.default}
\title{A dataset-specific method to overridden that returns default sample grouping}
\usage{
facet_frame(x, ...)

\method{facet_frame}{default}(x, ...)
}
\description{
A dataset-specific method to overridden that returns default sample grouping
}
