% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{FacileData-package}
\alias{FacileData}
\alias{FacileData-package}
\title{FacileData: A fluent API for accessing multi-assay high-throughput genomics data}
\description{
Defines the "FacileData" API over multi-assay, high-throughput
  genomics data. The FacileData API is a fluent grammar that facilitates
  exploratory data analysis. This package also defines a "FacileDataSet" class,
  which is a reference implementation of the "FacileData" API that uses SQLite
  and HDF5 files to store arbitrarily large datasets and provide fast and
  efficient access to arbitrary subsets of these data without loading it all
  into memory.
}
\author{
\strong{Maintainer}: Steve Lianoglou \email{lianoglou@dnli.com} (0000-0002-0924-1754)

Authors:
\itemize{
  \item Vincent Rouilly \email{rouilly.vincent@gene.com}
  \item Peter Haverty \email{haverty.peter@gene.com}
}

Other contributors:
\itemize{
  \item Denali Therapeutics (Coypright 2019) [copyright holder, funder]
  \item Genentech (Coypright 2016 - 2018) [copyright holder, funder]
}

}
