% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc-assay-containers.R
\name{as.FacileDataSet}
\alias{as.FacileDataSet}
\title{Converts Bioc assay containers into a FacileDataSet}
\usage{
as.FacileDataSet(x, path, ogranism, assays = NULL, metayaml = NULL, ...)
}
\value{
a \code{\link{FacileDataSet}}
}
\description{
Converts Bioc assay containers into a FacileDataSet
}
