% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FacileDataSet.R
\name{FacileDataSet}
\alias{FacileDataSet}
\title{Instantiates a FacileDataSet object from disk.}
\usage{
FacileDataSet(path, data.fn = file.path(path, "data.sqlite"),
  sqlite.fn = file.path(path, "data.sqlite"), hdf5.fn = file.path(path,
  "data.h5"), meta.fn = file.path(path, "meta.yaml"),
  anno.dir = file.path(path, "custom-annotation"), cache_size = 80000,
  db.loc = c("reference", "temporary", "memory"), ...)
}
\arguments{
\item{path}{The path to the FacileData repository}

\item{data.fn}{A custom path to the database (probably don't mess with this)}

\item{anno.dir}{A directory to house custom annotations/sample covariates}

\item{cache_size}{A custom paramter for the SQLite database}

\item{covdef.fn}{A custom path to the yaml file that has covariate mapping info}
}
\description{
The FacileDataSet is a reference data storage implementation that implements
the \strong{FacileData Access API}. It facilitates the storage and retrieval of
large amounts of data by leveraging a SQLite database to store sample- and
feature-level metadata ("\code{pData}" and "\code{fData}"), and an HDF5 file to store
all of the dense assay (matrix) data (gene counts, microarray intensities,
etc.).
}
\details{
TODO: Document the "FacileData Access API"
}
