% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{fds}
\alias{fds}
\alias{fds.FacileDataStore}
\alias{fds.default}
\alias{fds<-}
\alias{fds<-.tbl}
\alias{fds<-.data.frame}
\alias{set_fds}
\title{Get/set the FacileDataStore from an object}
\usage{
fds(x, ...)

\method{fds}{FacileDataStore}(x)

\method{fds}{default}(x, ...)

fds(x) <- value

\method{fds}{tbl}(x) <- value

\method{fds}{data.frame}(x) <- value

set_fds(x, value)
}
\arguments{
\item{x}{the object}

\item{db}{The \code{FacileDb} object}
}
\description{
Get/set the FacileDataStore from an object
}
