% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{fetch_assay_score}
\alias{fetch_assay_score}
\title{NOTE: is fetch_assay_score really necessary?}
\usage{
fetch_assay_score(x, features, samples = NULL, assay_name = NULL,
  as.matrix = FALSE, ..., subset.threshold = 700)
}
\description{
NOTE: is fetch_assay_score really necessary?
}
\seealso{
Other FacileInterface: \code{\link{facet_frame.FacileDataSet}},
  \code{\link{fetch_sample_covariates}},
  \code{\link{fetch_sample_statistics}},
  \code{\link{with_sample_covariates}}
}
\concept{FacileInterface}
