% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{fetch_expression}
\alias{fetch_expression}
\title{Helper function creates dplyr query to get expression data of interest.}
\usage{
fetch_expression(samples = NULL, feature_ids = NULL, db = FacileDb())
}
\arguments{
\item{db}{The db connection to use. Creates and destroys a new one
if not passed in.}

\item{indication}{The indication to get data from}

\item{subtype}{The subtype from the desired \code{indication}}

\item{sample_ids}{The sample_id's (barcodes) to fetch. If specified
this trumps whatever is specified in \code{indication} and
\code{subtype}}

\item{entrez_ids}{Restricts the fetch to only these genes}
}
\value{
A lazy \code{\link[dplyr]{tbl}} object with the expression
  data to be \code{\link[dplyr]{collect}}ed when \code{db} is provided,
  othwerise a \code{tbl_df} of the results.
}
\description{
Note that if \code{db} is not provided, the result will be \code{collect}ed,
because the database connection will be closed when the function exits.
}

