% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-covariates.R
\name{fetch_sample_covariates}
\alias{fetch_sample_covariates}
\title{Fetch rows from sample_covariate table for specified samples and covariates}
\usage{
fetch_sample_covariates(db, samples = NULL, covariates = NULL)
}
\arguments{
\item{db}{a \code{FacileDb} connection}

\item{samples}{a samples descriptor \code{tbl_*}}

\item{covariates}{character vector of covariate names}

\item{do.collect}{for collection of result from database?}
}
\value{
rows from the \code{sample_covaraite} table
}
\description{
Fetch rows from sample_covariate table for specified samples and covariates
}

