% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{fetch_sample_statistics}
\alias{fetch_sample_statistics}
\title{NOTE: fetch_sample_statistics -> \code{fetch_assay_covariates}
Issue #2}
\usage{
fetch_sample_statistics(x, samples = NULL, semi = TRUE, assay_name = NULL, ...)
}
\description{
NOTE: fetch_sample_statistics -> \code{fetch_assay_covariates}
Issue #2
}
\seealso{
Other FacileInterface: 
\code{\link{facet_frame.FacileDataSet}()},
\code{\link{fetch_assay_score}()},
\code{\link{fetch_sample_covariates}()},
\code{\link{with_sample_covariates}()}
}
\concept{FacileInterface}
