% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{fetch_samples}
\alias{fetch_samples}
\title{Fetches a sample descriptor that matches filter criterion over covariates.}
\usage{
fetch_samples(x, ...)
}
\arguments{
\item{x}{A \code{FacileDataSet} object}

\item{...}{filter clause to apply to \code{sample_covariate_tbl(x)}}
}
\value{
a facile sample descriptor with a \code{FacileDataSet} connection.
}
\description{
Fetches a sample descriptor that matches filter criterion over covariates.
}
\examples{
exampleFacileDataSet() \%>\%
  fetch_samples(indication \%in\% c('BRCA', 'COAD'))
}

