% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{fetch_samples}
\alias{fetch_samples}
\title{Fetches a sample descriptor that matches the filter criterion.}
\usage{
fetch_samples(x, ...)
}
\arguments{
\item{x}{A \code{FacileDataRepository}}

\item{...}{the NSE boolean filter criteria}
}
\value{
a facile sample descriptor
}
\description{
Use \code{...} as if this is a dplyr::filter call, and our
sample_covariate_tbl was "wide".
}
\details{
This is experimental, so each "term" in the filter criteria should be
just one boolean operation. Multiple terms passed into \code{...} will be
"AND"ed together.
}
