% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSE-filter-features.R
\name{filter_features}
\alias{filter_features}
\title{Filter against the sample_covariate_tbl as if it were wide.}
\usage{
filter_features(x, ...)
}
\arguments{
\item{x}{A \code{FacileDataSet}}
}
\description{
This feature is implemented so poorly. It's only really meant to be used
interactively, and with extreme caution ... programatically specifcying
column names in feature table, for instance, does not work right now.
}
\details{
TODO: Professionaly implement this.
}
