% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSE-filter-features.R
\name{filter_features}
\alias{filter_features}
\title{Filter against the sample_covariate_tbl as if it were wide.}
\usage{
filter_features(x, ...)
}
\arguments{
\item{x}{A \code{FacileDataSet}}

\item{...}{NSE claused to use in \code{\link[dplyr]{filter}} expressions}
}
\description{
This feature is implemented so poorly. It's only really meant to be used
interactively, and with extreme caution ... programatically specifying
column names in feature table, for instance, does not work right now.
}
\details{
TODO: Use tidyeval
}
\seealso{
Other API: \code{\link{facet_frame}},
  \code{\link{fetch_assay_data}},
  \code{\link{fetch_assay_score}},
  \code{\link{fetch_custom_sample_covariates}},
  \code{\link{fetch_sample_covariates}},
  \code{\link{fetch_sample_statistics}},
  \code{\link{fetch_samples}},
  \code{\link{filter_samples}}, \code{\link{organism}},
  \code{\link{samples}}
}
\concept{API}
