% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSE-filter-samples.R
\name{filter_samples}
\alias{filter_samples}
\title{Filter against the sample_covariate_tbl EAV table as if it were wide.}
\usage{
filter_samples(x, ..., with_covariates = FALSE)
}
\arguments{
\item{x}{A `FacileDataSet`}

\item{...}{NSE claused to use in [dplyr::filter()] expressions}
}
\value{
a sample-descriptor `data.frame` that includes the dataset,sample_id
  pairs that match the virtual `filter(covaries, ...)` clause executed here.
}
\description{
This allows the user to query the `FacileDataSet` as if it were a wide
`pData` `data.frame` of all its covariates.
}
\details{
This feature is only really meant to be
used interactively, and with extreme caution ... programatically specifying
the covariates, for instance, does not work right now.

TODO: Implement using `tidyeval`
}
\examples{
fds <- exampleFacileDataSet()

# To identify all samples that are of "CMS3" or "CMS4" subtype(
# stored in the "subtype_crc_cms" covariate:
crc.34 <- filter_samples(fds, subtype_crc_cms \%in\% c("CMS3", "CMS4"))
eav.query <- sample_covariate_tbl(fds) \%>\%
  filter(variable == "subtype_crc_cms", value \%in\% c("CMS3", "CMS4")) \%>\%
  collect
setequal(crc.34$sample_id, eav.query$sample_id)
}
\seealso{
Other API: \code{\link{facet_frame}},
  \code{\link{fetch_assay_data}},
  \code{\link{fetch_assay_score}},
  \code{\link{fetch_custom_sample_covariates}},
  \code{\link{fetch_sample_covariates}},
  \code{\link{fetch_sample_statistics}},
  \code{\link{fetch_samples}},
  \code{\link{filter_features}}, \code{\link{organism}},
  \code{\link{samples}}
}
\concept{API}
