% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{filter_samples}
\alias{filter_samples}
\title{Filters the samples down in a dataset to ones specified}
\usage{
filter_samples(x, samples = NULL, do.semi = TRUE)
}
\arguments{
\item{x}{likely a \code{tbl_sqlite} object, but a \code{tbl_df}-like
object should work as well.}

\item{samples}{a sample descriptor \code{tbl_df}-like object (likely a
\code{tbl_sqlite} object) that has \code{"dataset"} and \code{"samle_id"}
columns.}
}
\value{
filtered version of \code{x} that only has the desired samples
}
\description{
Tables like \code{expression} and \code{sample_covariate} house different
datapoints per sample, and we often want to only retreive data points over
a subset of samples.
}

