% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NSE-filter-samples.R
\name{filter_samples}
\alias{filter_samples}
\title{Filter against the sample_covariate_tbl as if it were wide.}
\usage{
filter_samples(x, ..., with_covariates = FALSE)
}
\arguments{
\item{x}{A \code{FacileDataSet}}

\item{...}{NSE claused to use in \code{\link[dplyr]{filter}} expressions}
}
\description{
This feature is implemented so poorly. It's only really meant to be used
interactively, and with extreme caution ... programatically specifcying
the covariates, for instance, does not work right now.
}
\details{
TODO: Professionaly implement this.
}
