% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FacileDataSet.R
\name{fds}
\alias{fds}
\alias{fds<-}
\alias{fds<-.data.frame}
\alias{fds<-.tbl}
\alias{set_fds}
\title{Get/set db}
\usage{
fds(x)

fds(x) <- value

\method{fds}{tbl}(x) <- value

\method{fds}{data.frame}(x) <- value

set_fds(x, value)
}
\arguments{
\item{x}{the object}

\item{db}{The \code{FacileDb} object}
}
\description{
Get/set db
}

