% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FacileDataSet.R
\name{hdf5fn}
\alias{hdf5fn}
\title{Get location of the FacileDataSet HDF5 file}
\usage{
hdf5fn(x, mustWork = TRUE)
}
\arguments{
\item{x}{FacileDataSet}

\item{mustWork}{single logical}
}
\value{
path to HDF5 file
}
\description{
Get location of the FacileDataSet HDF5 file
}
\seealso{
Other FacileDataSet: \code{\link{FacileDataSet}},
  \code{\link{dbfn}}, \code{\link{meta_file}}
}
\concept{FacileDataSet}
