% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construction.R
\name{initializeFacileDataSet}
\alias{initializeFacileDataSet}
\title{Create an empty FacileDataSet}
\usage{
initializeFacileDataSet(path, meta_file, page_size = 2^12,
  cache_size = 2e+05)
}
\arguments{
\item{path}{the directory to create which will house the
\code{FacileDataSet}}

\item{page_size, cache_size}{\code{pragma} values to setup the backend SQLite
database}

\item{covariate_definition}{the path to the covariate definition file}
}
\value{
inivisibly returns the path to the successfully created datastore
}
\description{
Create an empty FacileDataSet
}
