% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facile_interface_generics.R
\name{name}
\alias{name}
\title{Labeled acts like interface to reactive modules.}
\usage{
name(x, ...)
}
\description{
Modules that implement this interface must return `label` and `name` reactive
elements within them.
}
\details{
We use these when something (like a `assayFeatureSelect`) needs
a "computer friendly" name for itself (`name()`), or a more human readable
name (`label()`)
}
