% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mcast}
\alias{mcast}
\title{Cast a data.frame into a matrix.}
\usage{
mcast(data, formula, fun.aggregate = NULL, ...)
}
\arguments{
\item{data}{a \code{data.frame} to whip into a matrix}

\item{formula}{the pivot formula}

\item{fun.aggregate}{the aggregate function}

\item{...}{args past down to \code{\link[reshape2]{dcast}}.}
}
\description{
This is a simple wrapper to \code{\link[reshape2]{dcast}} which converts
the \code{data.frame} that call would create into a matrix which uses the
first column as the rownames.
}
