% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FacileDataSet.R
\name{meta_file}
\alias{meta_file}
\alias{meta_info}
\alias{dataset_definitions}
\title{Path to the meta information file}
\usage{
meta_file(x)

meta_info(x)

dataset_definitions(x, as.list = TRUE)
}
\arguments{
\item{x}{\code{FacileDataSet}}

\item{x}{\code{FacileDataSet}}
}
\description{
Path to the meta information file

Get meta information for dataset
}
