% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facile_interface_generics.R
\name{organism}
\alias{organism}
\title{The Facile API}
\usage{
organism(x, ...)
}
\description{
This is a stub for a manpage all about the FacileAPI
}
\seealso{
Other FacileInterface: \code{\link{covariate_summary.FacileDataSet}},
  \code{\link{facet_frame.FacileDataSet}},
  \code{\link{sample_covariates}}
}
\concept{FacileInterface}
