% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.FacileDataSet.R
\name{pdata}
\alias{pdata}
\title{Bioc-container specific pData extraction functions}
\usage{
pdata(x, covariate_metadata = NULL, ...)
}
\arguments{
\item{x}{SummarizedExperiment, ExpressionSet or DGEList}

\item{...}{additional args, ignored for now}
}
\description{
This is an internal function, but exported so it is registered and found
post R 4.0
}
