% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facile_interface_generics.R,
%   R/sample-covariates.R
\name{fetch_sample_covariates}
\alias{fetch_sample_covariates}
\alias{fetch_sample_covariates.FacileDataSet}
\alias{fetch_sample_covariates.facile_frame}
\title{Fetch rows from sample_covariate table for specified samples and covariates}
\usage{
fetch_sample_covariates(x, samples = NULL, covariates = NULL,
  custom_key = Sys.getenv("USER"), with_source = FALSE, ...)

\method{fetch_sample_covariates}{FacileDataSet}(x, samples = NULL,
  covariates = NULL, custom_key = Sys.getenv("USER"),
  with_source = FALSE, ...)

\method{fetch_sample_covariates}{facile_frame}(x, samples = NULL,
  covariates = NULL, custom_key = Sys.getenv("USER"),
  with_source = FALSE, ...)
}
\arguments{
\item{x}{a \code{FacileDataSet} connection}

\item{samples}{a samples descriptor \code{tbl_*}}

\item{covariates}{character vector of covariate names}

\item{custom_key}{The key to use to fetch more custom annotations over
the given samples}
}
\value{
rows from the \code{sample_covariate} table
}
\description{
Fetch rows from sample_covariate table for specified samples and covariates
}
\seealso{
Other FacileInterface: \code{\link{facet_frame.FacileDataSet}},
  \code{\link{fetch_assay_score}},
  \code{\link{fetch_sample_statistics}},
  \code{\link{with_sample_covariates}}

Other API: \code{\link{fetch_assay_data.FacileDataSet}},
  \code{\link{fetch_assay_score.FacileDataSet}},
  \code{\link{fetch_custom_sample_covariates.FacileDataSet}},
  \code{\link{fetch_sample_statistics.FacileDataSet}},
  \code{\link{fetch_samples.FacileDataSet}},
  \code{\link{filter_features.FacileDataSet}},
  \code{\link{filter_samples.FacileDataSet}},
  \code{\link{organism.FacileDataSet}},
  \code{\link{samples.FacileDataSet}}

Other API: \code{\link{fetch_assay_data.FacileDataSet}},
  \code{\link{fetch_assay_score.FacileDataSet}},
  \code{\link{fetch_custom_sample_covariates.FacileDataSet}},
  \code{\link{fetch_sample_statistics.FacileDataSet}},
  \code{\link{fetch_samples.FacileDataSet}},
  \code{\link{filter_features.FacileDataSet}},
  \code{\link{filter_samples.FacileDataSet}},
  \code{\link{organism.FacileDataSet}},
  \code{\link{samples.FacileDataSet}}

Other API: \code{\link{fetch_assay_data.FacileDataSet}},
  \code{\link{fetch_assay_score.FacileDataSet}},
  \code{\link{fetch_custom_sample_covariates.FacileDataSet}},
  \code{\link{fetch_sample_statistics.FacileDataSet}},
  \code{\link{fetch_samples.FacileDataSet}},
  \code{\link{filter_features.FacileDataSet}},
  \code{\link{filter_samples.FacileDataSet}},
  \code{\link{organism.FacileDataSet}},
  \code{\link{samples.FacileDataSet}}
}
\concept{API}
\concept{FacileInterface}
