% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facile_interface_generics.R,
%   R/sample-covariates.R
\name{sample_covariates}
\alias{sample_covariates}
\alias{sample_covariates.FacileDataSet}
\title{Fetches summary information about sample covariates}
\usage{
sample_covariates(x, samples = NULL, ...)

\method{sample_covariates}{FacileDataSet}(x, samples = active_samples(x),
  ...)
}
\arguments{
\item{x}{a \code{FacileDataSet} connection}

\item{samples}{a samples descriptor, used to narrow the universe of samples
to get sample covariate information from.}

\item{...}{extras}
}
\value{
a tibble of sample covariate information
}
\description{
Returns a table with variable,class,nsamples columns
}
\seealso{
Other FacileInterface: \code{\link{facet_frame.FacileDataSet}},
  \code{\link{organism}}
}
\concept{FacileInterface}
