% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FacileDataSet.R
\name{samples}
\alias{samples}
\title{Get tibble of sample information}
\usage{
samples(x)
}
\arguments{
\item{x}{FacileDataSet}
}
\value{
tibble of sample attributes
}
\description{
Get tibble of sample information
}
