% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{samples}
\alias{samples}
\title{Returns a facile_frame of samples from an object in the faciebio ecosystem.}
\usage{
samples(x, ...)
}
\description{
\code{samples} can be extracted from a FacileDataStore itself, or the result of
an analysis initiated from a FacileDataStore (like you'll find in the
FacileAnalysis package).
}
\seealso{
Other FacileInterface: 
\code{\link{facet_frame.FacileDataSet}()},
\code{\link{fetch_assay_score}()},
\code{\link{fetch_sample_covariates}()},
\code{\link{fetch_sample_statistics}()},
\code{\link{with_assay_data}()}
}
\concept{FacileInterface}
