% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-covariates.R
\name{save_custom_sample_covariates}
\alias{save_custom_sample_covariates}
\title{Saves custom sample covariates to a FacileDataSet}
\usage{
save_custom_sample_covariates(x, name, annotation, custom_key = "anonymous",
  file.prefix = "facile", sample_filter_critera = NULL)
}
\arguments{
\item{x}{the \code{FacileDataSet}}

\item{name}{the variable name of the covariate}

\item{annotation}{the annotation table of covariate vaues to a
sample-descriptor-like table}

\item{custom_key}{the custom key (likely userid) for the annotation}

\item{file.prefix}{Vincent uses this}

\item{sample_filter_criteria}{optional list of filtering criteria that were
used to drill down into the samples we have the \code{annotatino}
data.frame for}
}
\description{
Saves custom sample covariates to a FacileDataSet
}

