% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-covariates.R
\name{spread_covariates}
\alias{spread_covariates}
\title{Spreads the covariates returned from database into wide data.frame}
\usage{
spread_covariates(x, .fds = fds(x), cov.def = NULL, ...)
}
\arguments{
\item{x}{output from \code{fetch_sample_covariates}}

\item{.fds}{A \code{FacileDataSet} object}
}
\value{
a wide \code{tbl_df}-like object
}
\description{
Samples that did not have a value for a specific covariate are assigned to
have NA.
}
