% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{example_sample_covariates}
\alias{example_sample_covariates}
\alias{example_meta}
\alias{example_sample_covariate_definitions}
\title{Fetches exemplar pData}
\usage{
example_sample_covariates()

example_meta(file.path = FALSE)

example_sample_covariate_definitions()
}
\arguments{
\item{file.path}{If `TRUE`, returns the path to the yaml file, otherwise
returns the list-of-list meta definition.}
}
\value{
Either the list-of-list meta definition, or path to the `meta.yaml`
  file where these are defined.

the list-of-list definitions for the example `pData` returned from
  [example_sample_covariates()]
}
\description{
Fetches exemplar pData

Fetches the meta file for the example FacileDataSet

Fetches the EAV definitions for the sample covariates
}
