% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity-attribute-value.R
\name{validate_covariate_def_list}
\alias{validate_covariate_def_list}
\title{Validates that a covariate defintion list reasonably describes a data.frame.}
\usage{
validate_covariate_def_list(x, pdata)
}
\arguments{
\item{x}{a covariate definition list-of-lists}

\item{pdata}{a \code{data.frame}}
}
\description{
The covariates defined in \code{x} must be a subset of the columns in \code{pdata}.
This method will throw an error if there is a covariate in \code{x} that does
not have a matching column in \code{pdata}.
}
\details{
This function does not check if all columns in \code{pdata} have definitions in
\code{x}.
}
