% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{with_assay_data}
\alias{with_assay_data}
\title{Append assay data to a facile_frame}
\usage{
with_assay_data(
  x,
  features,
  assay_name = NULL,
  normalized = TRUE,
  aggregate = FALSE,
  aggregate.by = "ewm",
  spread = TRUE,
  with_assay_name = FALSE,
  ...,
  verbose = FALSE,
  .fds = NULL
)
}
\description{
Append assay data to a facile_frame
}
\seealso{
Other FacileInterface: 
\code{\link{facet_frame.FacileDataSet}()},
\code{\link{fetch_assay_score}()},
\code{\link{fetch_sample_covariates}()},
\code{\link{fetch_sample_statistics}()},
\code{\link{samples}()}
}
\concept{FacileInterface}
