% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay-data.R, R/expression.R
\name{with_expression}
\alias{with_expression}
\alias{with_expression}
\title{Append expression values to sample-descriptor}
\usage{
with_expression(samples, feature_ids, with_symbols = TRUE,
  .fds = fds(samples))

with_expression(samples, feature_ids, with_symbols = TRUE,
  .fds = fds(samples))
}
\arguments{
\item{feature_ids}{character vector of feature_ids}

\item{with_symbols}{Do you want gene symbols returned, too?}

\item{.fds}{A \code{FacileDataSet} object}

\item{x}{a samples descriptor}

\item{x}{a samples descriptor}

\item{feature_ids}{character vector of feature_ids}

\item{with_symbols}{Do you want gene symbols returned, too?}

\item{.fds}{A \code{FacileDataSet} object}
}
\value{
a tbl-like result

a tbl-like result
}
\description{
Append expression values to sample-descriptor

Append expression values to sample-descriptor
}
