% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-covariates.R
\name{with_sample_covariates}
\alias{with_sample_covariates}
\title{Appends covariate columns to a query result}
\usage{
with_sample_covariates(x, covariates = NULL, .fds = fds(x))
}
\arguments{
\item{x}{a facile sample descriptor}

\item{covariates}{character vector of covariate names}

\item{.fds}{A \code{FacileDataSet} object}
}
\value{
The facile \code{x} object, annotated with the specified covariates.
}
\description{
Note that this function will force the collection of \code{x}
}

