% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facile_interface_generics.R
\name{with_sample_covariates}
\alias{with_sample_covariates}
\title{Appends covariate columns to a query result}
\usage{
with_sample_covariates(x, covariates = NULL, na.rm = FALSE,
  custom_key = Sys.getenv("USER"), .fds = NULL, ...)
}
\arguments{
\item{x}{a facile sample descriptor}

\item{covariates}{character vector of covariate names. If \code{NULL}
(default), returns all covariates, if is character and length() == 0, then
this is a no-op (x is returned)}

\item{na.rm}{if \code{TRUE}, filters outgoing result such that only rows
with nonNA values for the \code{covariates} specified here will be
returned. Default: \code{FALSE}. Note that this will not check columns
not specified in \code{covariates} for NA-ness.}

\item{custom_key}{The key to use to fetch more custom annotations over
the given samples}

\item{.fds}{A \code{FacileDataSet} object}
}
\value{
The facile \code{x} object, annotated with the specified covariates.
}
\description{
Note that this function will force the collection of \code{x}
}
\seealso{
Other FacileInterface: \code{\link{facet_frame.FacileDataSet}},
  \code{\link{fetch_assay_score}},
  \code{\link{fetch_sample_covariates}},
  \code{\link{fetch_sample_statistics}}
}
\concept{FacileInterface}
