% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-covariates.R
\name{with_sample_covariates}
\alias{with_sample_covariates}
\title{Appends covariate columns to a query result}
\usage{
with_sample_covariates(x, covariates = NULL, db = attr(x, "db"),
  cov.def = db[["cov.def"]])
}
\description{
Note that this function will force the collection of \code{x}
}

